% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkQuality.R
\name{checkQuality}
\alias{checkQuality}
\title{Quality control step for .fastq files}
\usage{
checkQuality(files)
}
\arguments{
\item{files}{The local directory to sRNA-seq sample. These can be raw or
trimmed.}
}
\value{
A "qc" folder within the directory containing \code{fastqc} quality
check files for each sample in \code{HTML} format.
}
\description{
Invokes an OS command to run unzip any compressed .gz files
and run a quality control analysis using \code{fastqc} on the files within the
directory.
}
\examples{
\dontrun{
raw_samples <- "/Users/user1/projectname/workplace/raw/"
qc <- checkQuality(raw_samples)
}
}
