\name{chr12_Eggplant_V4.1.fa.gz}
\alias{chr12_Eggplant_V4.1.fa.gz}
\docType{data}
\title{
chr12_Eggplant_V4.1.fa.gz}
\description{
Chromosome 12 of genome assebly Solanium melongena Version 4.1
}
\usage{data("inst/extdata/chr12_Eggplant_V4.1.fa.gz")}
\format{
  FASTA gzip
  \describe{
  }
}
\details{
}
\source{
Barchi, L., Rabanus‐Wallace, M. T., Prohens, J., Toppino, L., Padmarasu, S.,
Portis, E., ... & Giuliano, G. (2021). Improved genome assembly and pan‐genome
provide key insights into eggplant domestication and breeding.
The Plant Journal, 107(2), 579-596.

}
\references{
https://solgenomics.net/organism/Solanum_melongena/genome
}
\examples{
chr12_Eggplant_V4.1 <- system.file("extdata",
                                   "chr12_Eggplant_V4.1.fa.gz",
                                    package="mobileRNA")
}
\keyword{datasets}
