% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr12_Eggplant_V4.1_function_IPR_final.gff.gz}
\alias{chr12_Eggplant_V4.1_function_IPR_final.gff.gz}
\title{chr12_Eggplant_V4.1_function_IPR_final.gff.gz}
\format{
GFF gzip
}
\source{
Barchi, L., Rabanus‐Wallace, M. T., Prohens, J., Toppino, L., Padmarasu, S.,
Portis, E., ... & Giuliano, G. (2021). Improved genome assembly and pan‐genome
provide key insights into eggplant domestication and breeding.
The Plant Journal, 107(2), 579-596.
}
\usage{
system.file("extdata",
"chr12_Eggplant_V4.1_function_IPR_final.gff.gz", package="mobileRNA")
}
\description{
Chromosome 12 of genome annotation Solanium melongena
Version 4.1
}
\examples{
system.file("extdata",
"chr12_Eggplant_V4.1_function_IPR_final.gff.gz", package="mobileRNA")
}
\references{
https://solgenomics.net/organism/Solanum_melongena/genome
}
\keyword{dataset}
