% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr2_ITAG4.0_gene_models.gff.gz}
\alias{chr2_ITAG4.0_gene_models.gff.gz}
\title{chr2_ITAG4.0_gene_models.gff.gz}
\format{
GFF gzip
}
\usage{
system.file("extdata",
"chr2_ITAG4.0_gene_models.gff.gz", package="mobileRNA")
}
\description{
Chromosome 2 of genome annotation Solanium lycopersicum (Heinz)
Version SL4.0
}
\examples{
system.file("extdata",
"chr2_ITAG4.0_gene_models.gff.gz", package="mobileRNA")
}
\references{
https://solgenomics.net/organism/Solanum_lycopersicum/genome
}
\keyword{dataset}
