\name{chr2_ITAG4.0_gene_models.gff.gz}
\alias{chr2_ITAG4.0_gene_models.gff.gz}
\docType{data}
\title{
chr2_ITAG4.0_gene_models.gff.gz
}
\description{
Chromosome 2 of genome annotation Solanium lycopersicum (Heinz) Version SL4.0
}
\usage{data("chr2_ITAG4.0_gene_models.gff.gz")}
\format{
GFF gzip
\describe{
  }
}
\details{
}
\source{
https://solgenomics.net/organism/Solanum_lycopersicum/genome
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
chr2_S_lycopersicum <- system.file("extdata",
                                   "chr2_ITAG4.0_gene_models.gff.gz",
                                    package="mobileRNA")
}
\keyword{datasets}
