% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr2_S_lycopersicum_chromosomes.4.00.fa.gz}
\alias{chr2_S_lycopersicum_chromosomes.4.00.fa.gz}
\title{chr2_S_lycopersicum_chromosomes.4.00.fa.gz}
\format{
FASTA gzip
}
\source{
Hosmani, P. S., Flores-Gonzalez, M., van de Geest, H., Maumus, F., Bakker, L.
V., Schijlen, E., ... & Saha, S. (2019). An improved de novo assembly and
annotation of the tomato reference genome using single-molecule sequencing,
Hi-C proximity ligation and optical maps. biorxiv, 767764.
}
\usage{
system.file("extdata",
"chr2_S_lycopersicum_chromosomes.4.00.fa.gz", package="mobileRNA")
}
\description{
Chromosome 2 of genome assembly Solanium lycopersicum (Heinz)
Version SL4.0
}
\examples{
system.file("extdata",
"chr2_S_lycopersicum_chromosomes.4.00.fa.gz", package="mobileRNA")
}
\references{
https://solgenomics.net/organism/Solanum_lycopersicum/genome
}
\keyword{dataset}
