\name{chr2_S_lycopersicum_chromosomes.4.00.fa.gz}
\alias{chr2_S_lycopersicum_chromosomes.4.00.fa.gz}
\docType{data}
\title{
chr2_S_lycopersicum_chromosomes.4.00.fa.gz
}
\description{
Chromosome 2 of genome assebly Solanium lycopersicum (Heinz) Version SL4.0

}
\usage{data("inst/extdata/chr2_S_lycopersicum_chromosomes.4.00.fa.gz")}
\format{
  FASTA gzip
}
\details{
}
\source{
Hosmani, P. S., Flores-Gonzalez, M., van de Geest, H., Maumus, F., Bakker, L.
    V., Schijlen, E., ... & Saha, S. (2019). An improved de novo assembly and
    annotation of the tomato reference genome using single-molecule sequencing,
    Hi-C proximity ligation and optical maps. biorxiv, 767764.}
\references{
https://solgenomics.net/organism/Solanum_lycopersicum/genome
}
\examples{

chr2_S_lycopersicum <- system.file("extdata",
                                   "chr2_S_lycopersicum_chromosomes.4.00.fa.gz",
                                    package="mobileRNA")
}
\keyword{datasets}
