% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_modify.R
\name{chr_modify}
\alias{chr_modify}
\title{Remove punctuation from chromosome names.}
\usage{
chr_modify(old, new, genome)
}
\arguments{
\item{old}{Character string of chromosome name to change.}

\item{new}{Replacement of old string. A character string to rename the
chromosomes.}

\item{genome}{Directory to the location of the merged genome reference file.}
}
\value{
Edited ref_merged.fasta file representing the merged genome
reference, saved in the "cat" directory within "reference" directory.
}
\description{
Invokes an OS command to alter the names of chromosomes by
selecting a character sting within the chromosomes names to replace.
See [funs(chr_checker)] to identify the names of the chromosomes in a genome.
}
\examples{
\dontrun{
# The Solanum lycopersicum (version 4) genome contains a
# full-stop/period within each chromosome name which needs to be removed.
# Each chromosome name starts with "SL4." and we will be changing it to
# remove the full-stop.


old <- "SL4."
new <- "SL4"
genome <- "/Users/user1/projectname/workplace/reference/cat/ref_merged.fa"
chr_punct(old, new, genome)
}

}
