% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrInfo.R
\name{chrInfo}
\alias{chrInfo}
\title{Calculate chromosome sizes, and store in a text file.}
\usage{
chrInfo(genome, out)
}
\arguments{
\item{genome}{Directory path to the genome (\code{.fa/.fasta})}

\item{out}{Path to output directory}
}
\value{
A plain text file containing the chromosome sizes
}
\description{
Invokes an OS command to calculate the chromosome sizes within
a genome.
}
\examples{
\dontrun{
ref <- "/Users/user1/projectname/workplace/reference/cat/ref_merged.fa"
save <- "/Users/user1/projectname/workplace/reference/cat/"
chrInfo(ref)
}

}
