% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workplace.R
\name{create_workplace}
\alias{create_workplace}
\title{Make the working directory structure}
\usage{
create_workplace(dirwd)
}
\arguments{
\item{dirwd}{Working Directory path. This argument could include using
the [getwd()] function to retrieve path.}
}
\value{
A working directory with folders the folders `"raw"` to store raw
sequencing data (`.fastq/.fq`), `"reference"` to store reference
genomes (`.fa/.fasta`) & `"annotation"` to store reference
annotation files (`.gff/.gff3`)
}
\description{
Creates the basic directory structure to store your raw data,
genome references and annotation files
}
\examples{
\dontrun{
create_workplace(getwd())
}
}
