% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkplace.R
\name{createWorkplace}
\alias{createWorkplace}
\title{Create a working directory file structure}
\usage{
createWorkplace(wd)
}
\arguments{
\item{wd}{Path to working directory.}
}
\value{
A working directory with folders the folders `"raw"` to store raw
sequencing data (`.fastq/.fq`), `"reference"` to store reference
genomes (`.fa/.fasta`) & `"annotation"` to store reference
annotation files (`.gff/.gff3`)
}
\description{
Creates the basic directory to store your raw data,
genome references and annotation files
}
\details{
This argument could include using the [base::getwd()] function to
retrieve path to the working directory.
}
\examples{
\dontrun{
createWorkplace(getwd())
}
}
