% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE_analysis}
\alias{DE_analysis}
\title{Differential Expression (DE) Analysis using DESeq2 or edgeR}
\usage{
DE_analysis(data, groupPair, method = c("edgeR", "DESeq2"))
}
\arguments{
\item{data}{Columns containing the raw count data for each sample within
the dataframe object.}

\item{groupPair}{A vector of the two conditions in the comparison}

\item{method}{The method to calculate normalisation based on
the `DESeq2` or `edgeR` package.
Must be stated as either `"DESeq2"` or `"edgeR"`.}
}
\value{
The differential expression analysis results.
}
\description{
This function allows you to compute the differential expression
of siRNA clusters. The function allows the choice between analysis with
DESeq2 or edgeR. This should be chosen based on the normalisation method
(See function [DE_prepare()]).
}
\examples{
## Differential analysis: DESeq2 example (24-nt & 21/22-nt sRNA)
require(DESeq2)


# 24-nt sRNA data
data("sRNA_24_prep_DESeq2")
sRNA_24_DE_DESeq2 <- DE_analysis(data = sRNA_24_prep_DESeq2,
                           groupPair = c("Tomato/Eggplant","Tomato/Tomato"),
                           method = "DESeq2" )
# 21/22-nt sRNA data
data("sRNA_2122_prep_DESeq2")
sRNA_2122_DE_DESeq2 <- DE_analysis(data = sRNA_2122_prep_DESeq2,
                           groupPair = c("Tomato/Eggplant","Tomato/Tomato"),
                           method = "DESeq2" )



## Differential analysis edgeR example (24-nt & 21/22-nt sRNA)
require(edgeR)
# 24-nt sRNA data
data("sRNA_24_prep_edgeR")
sRNA_24_DE_edgeR <- DE_analysis(data = sRNA_24_prep_edgeR,
                   groupPair = c("Tomato/Eggplant","Tomato/Tomato"),
                     method = "edgeR" )
# 21/22-nt sRNA data
data("sRNA_2122_prep_edgeR")
sRNA_2122_DE_edgeR <- DE_analysis(data = sRNA_2122_prep_edgeR,
                           groupPair = c("Tomato/Eggplant","Tomato/Tomato"),
                           method = "edgeR" )





}
