% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_results.R
\name{DE_results}
\alias{DE_results}
\title{Extract Differential Expression (DE) analysis results and add to the summary
dataframe.}
\usage{
DE_results(Summarydata, analysisResults, method = c("edgeR", "DESeq2"))
}
\arguments{
\item{Summarydata}{Summary dataframe containing the collective information on the
sRNA clusters.}

\item{analysisResults}{Object contain the DE analysis results
(see [DE_prepare()] and [DE_analysis()] for prior steps in
DE analysis)}

\item{method}{The method used to calculate the DE analysis: either
`DESeq2` or  `edgeR`  method.}
}
\value{
Adds extra columns of data to the dataframe. A column containing the
Log2 Fold-Change results, called `log2FoldChange`, a column containing
the p-values, called `pvalue`, a column containing the p-adjusted value,
called `padjusted`, and if using the `edgeR` method, a column containing the
log counts per million, called `logCPM`, are added.
}
\description{
A function to add the differential expression result to the
dataframe containing all information on sRNA clusters. This function
dependents on which method was used to call the DE analysis.
}
\examples{
# load summary data
data('sRNA_24')
data('sRNA_2122')



## DESeq2 example (24-nt & 21/22-nt sRNA)
# 24nt subset
data('sRNA_24_DE_DESeq2')
res_sRNA_24_DESeq2 <- DE_results(Summarydata = sRNA_24,
                                 analysisResults = sRNA_24_DE_DESeq2 ,
                                 method = "DESeq2")

# 2122nt subset
data('sRNA_2122_DE_DESeq2')
res_sRNA_2122_DESeq2 <- DE_results(Summarydata = sRNA_2122,
                                 analysisResults = sRNA_2122_DE_DESeq2 ,
                                 method = "DESeq2")


## edgeR example (24-nt & 21/22-nt sRNA)
require(edgeR)

# 24nt subset
data('sRNA_24_DE_edgeR')
res_sRNA_24_edgeR <- DE_results(Summarydata = sRNA_24,
                                 analysisResults = sRNA_24_DE_edgeR,
                                 method = "edgeR")
data('sRNA_2122_DE_edgeR')
res_sRNA_2122_edgeR <- DE_results(Summarydata = sRNA_2122,
                                 analysisResults = sRNA_2122_DE_edgeR,
                                 method = "edgeR")


}
