% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_plot.R
\name{distance_plot}
\alias{distance_plot}
\title{Heatmap to access sample distance as a quality control step}
\usage{
distance_plot(data)
}
\arguments{
\item{data}{a `DESeqDataSet` or `DGEList` data object depending on
the method see [DE_prepare()] for more information.}
}
\value{
A blue scale heatmap illustrating the sample distance
}
\description{
A function to draw a simple hierarchical clustered heatmap to observe
sample distance.
}
\examples{

## DESeq2 example: Sample distance using a Heatmap

# plot 24-nt sRNA dataset
data("sRNA_24_prep_DESeq2")
p1 <- distance_plot(sRNA_24_prep_DESeq2)

# plot 21/22-nt sRNA dataset
data("sRNA_2122_prep_DESeq2")
p2 <- distance_plot(sRNA_2122_prep_DESeq2 )



\dontrun{
## edgeR example: Sample distance using a Heatmap

# plot 24-nt sRNA dataset
data("sRNA_24_prep_edgeR")
p3 <- distance_plot(sRNA_24_prep_edgeR)

# plot 21/22-nt sRNA dataset
data("sRNA_2122_prep_edgeR")
p4 <- distance_plot(sRNA_2122_prep_edgeR)
}

}
