% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{heterograft_demo_1.fq}
\alias{heterograft_demo_1.fq}
\title{heterograft_demo_1.fq}
\usage{
base::system.file("extdata/sRNAseq","heterograft_demo_1.fq",package="mobileRNA")
}
\value{
FASTQ file
}
\description{
Demo FASTQ file
}
\details{
Demo sRNAseq FASTQ file for a heterograft sample where the sample
was taken from the shoot of the scion. The scion is Eggplant
(Solanum melongena) and the rootstock is Tomato (Solanum lycopersicum).
This file has been majorly reduced and does not represent true data.
For pre-processing purposes.
}
\examples{
 heterograft_demo_1 <- base::system.file("extdata/sRNAseq","heterograft_demo_1.fq",
 package="mobileRNA")
}
\keyword{FASTQ}
