% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Merge the reference genomes into a single reference genome}
\usage{
merge_files(files, file1 = NULL, file2 = NULL, out)
}
\arguments{
\item{files}{Path to the directory containing genome both reference files
(The directory must only contain these files).}

\item{file1}{Path to the directory of the first genome reference file.}

\item{file2}{Path to the directory of the second genome reference file.}

\item{out}{Path desired output directory.}
}
\value{
A file named `"ref_merged.fa"` representing the merged
genome reference.
}
\description{
Merges two reference genomes (`.fa/.fasta)`. into one.
}
\examples{
\dontrun{
merge_files(file1 = "./workplace/reference/ref1.fa.gz",
            file2 = "./workplace/reference/ref2.fa.gz"
            out = "./workplace/reference/merge/")
# Or,
merge_files(files = "./workplace/reference/",
            out = "./workplace/reference/merge/")
}

}
