% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_molecules.R
\name{mobile_molecules}
\alias{mobile_molecules}
\title{Identify mobile sRNA molecules}
\usage{
mobile_molecules(
  data,
  controls,
  id,
  task = c("keep", "remove"),
  statistical = TRUE,
  padj = 0.05,
  p.value = NULL
)
}
\arguments{
\item{data}{Numeric data frame}

\item{controls}{Vector containing names of control samples.}

\item{id}{A string related to the chromosomes in a particular genome}

\item{task}{An option to keep or remove the chromosomes containing the
identifying string. To keep the chromosomes with the ID, set task=keep.
To remove, set `task="remove"`.}

\item{statistical}{If TRUE, will undertake statistical filtering based on the a
p-value or adjusted p-value threshold stated by `padj` and `p.value`.}

\item{padj}{A user defined numeric value to represent the adjusted p-value
threshold to define statistic significance. Defaults set at 0.05.Only mobile
molecules with adjusted p-values equal or lower than specified are returned.}

\item{p.value}{A user defined numeric value to represent the p-value
threshold to define statistic significance. There is no default value, set
this instead of using an adjusted p-value to filter molecules. Only mobile
molecules with p-values equal or lower than specified are returned.}
}
\value{
A data-frame containing statistically significant mobile small
RNA molecules.
}
\description{
A function to identify mobile small RNA molecules traveling
from one genotype to another in a hetero-grafted system.
}
\examples{


## DESeq2 example (24-nt & 21/22-nt sRNA)
data("res_sRNA_24_DESeq2")
data("res_sRNA_2122_DESeq2")


# vector of control names
controls <- c("TomTom_1", "TomTom_2", "TomTom_3")

# Identify mobile 24-nt sRNA from eggplant genome:
# remove clusters associated to tomato
sRNA_24_mobile_DEseq2 <- mobile_molecules(data = res_sRNA_24_DESeq2,
                                         controls = controls,
                                         id = "SL40",
                                         task = "remove")

# Identify mobile 21/22-nt sRNA from eggplant genome:
# remove clusters associated to tomato
sRNA_2122_mobile_DEseq2  <- mobile_molecules(data = res_sRNA_2122_DESeq2,
                                             controls = controls,
                                             id = "SL40",
                                             task = "remove")






## edgeR example (24-nt & 21/22-nt sRNA)

data("res_sRNA_24_edgeR")
data("res_sRNA_2122_edgeR")
# vector of control names
controls <- c("TomTom_1", "TomTom_2", "TomTom_3")


# Identify mobile 24-nt sRNA from eggplant genome:
# remove clusters associated to tomato
sRNA_24_mobile_edgeR <- mobile_molecules(data = res_sRNA_24_edgeR,
                                         controls = controls,
                                         id = "SL40",
                                         task = "remove")

# Identify mobile 21/22-nt sRNA from eggplant genome:
# remove clusters associated to tomato
sRNA_2122_mobile_edgeR <- mobile_molecules(data = res_sRNA_2122_edgeR,
                                           controls = controls,
                                           id = "SL40",
                                           task = "remove")

}
