% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{mRNA_data}
\alias{mRNA_data}
\title{mRNA_data: simulated messenger RNA data for biological replicates}
\usage{
data(mRNA_data)
}
\value{
Dataframe in global environment
}
\description{
Simulated mRNAseq dataset
}
\details{
Simulates data is taken from eggplant and tomato mRNAseq samples and
created to simulate to movement of mRNA molecules from an Tomato rootstock to
an Eggplant Scion. Two Eggplant replicates were spiked with the same 150
tomato mRNA clusters, and named "heterograft_" 1 to 2. The analysis compares
these heterografts to two Eggplant self-grafts which are the original
un-spiked Eggplant replicates, called "selfgraft_" 1 to 2.

This data was imported and organised by the \code{\link[=RNAimport]{RNAimport()}} function.
}
\examples{
 data("mRNA_data")
}
