% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNAimport.R
\name{mRNAimport}
\alias{mRNAimport}
\title{Import and organise mRNAseq data set}
\usage{
mRNAimport(directory, samples, report = TRUE, tidy = TRUE)
}
\arguments{
\item{directory}{Path to directory containing of sample folders.}

\item{samples}{Vector of characters naming the sample names correlating
to outputted folders located in the \code{`directory`} argument path.}

\item{report}{Logical; prompts progress.}

\item{tidy}{Logical; removes genes from analysis where there are zero counts
across all replicates.}
}
\description{
Organises mRNAseq counts data from each sample replicate into
a single dataframe.
}
\details{
Supply the directory to the folder storing the plain text files for each
sample replicate contain counts data. Each file is draw from the directory
based on the file names supplied to the samples argument, minus the .txt
extension. The file names supplied will represent the sample across the
analysis, unless changed or by altering column names after importation with
this function.

The counts data is composed of two columns; the first containing gene names,
and the second containing counts.

The first column of the function output contains all the gene names identified
across all samples. The other columns represent the counts for each sample
replicate imported. Where the sample name is incorporated into the column name.
For example, for a sample called "Sample1", the counts will be stored in a
column called "Count_Sample1".
}
\examples{
\dontrun{

df_mRNA <- RNAimport(input = "mRNA",
                     directory = "./analysis/mRNA_counts/",
                     samples = c( "TomEgg_1", "TomEgg_2", "TomEgg_3",
                                  "TomTom_1" , "TomTom_2" , "TomTom_3"))

}

}
