% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{PCA for quality control}
\usage{
PCA_plot(data)
}
\arguments{
\item{data}{Either a \code{DESeqDataSet} or \code{DGEList} containing
organised or normalised counts based on the method of analysis.}
}
\value{
A PCA plot to show sample distance.
}
\description{
A function to draw a PCA plot. The function undertakes rlog
 transformation of the data in an unbiased manner (\code{blind=TRUE}).
}
\examples{

## DESeq2 example: Sample distance using a PCA plot

# plot 24-nt sRNA dataset
 data("sRNA_24_prep_DESeq2")
p1 <- PCA_plot(sRNA_24_prep_DESeq2)


# plot 21/22-nt sRNA dataset
 data("sRNA_2122_prep_DESeq2")
p2 <- PCA_plot(sRNA_2122_prep_DESeq2 )



\dontrun{
## edgeR example: Sample distance using a PCA plot

# plot 24-nt sRNA dataset
data("sRNA_24_prep_edgeR")
p3 <- PCA_plot(sRNA_24_prep_edgeR)

# plot 21/22-nt sRNA dataset
data("sRNA_2122_prep_edgeR")
p4 <- PCA_plot(sRNA_2122_prep_edgeR)
}






}
