% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsensusFrequency.R
\name{plotConsensusFrequency}
\alias{plotConsensusFrequency}
\title{Plot Consensus small RNA frequency}
\usage{
plotConsensusFrequency(data, relative = FALSE)
}
\arguments{
\item{data}{A data-frame (See [sample_table()] and
[define_consensus()] to produce an appropriate data-frame).
Plots data based on column labeled `"sRNA_Consensus"`.}

\item{relative}{If `relative=TRUE`, will plot the relative frequency.
As default, the function will plot the absolute frequency
(`relative == FALSE`) rather than the relative frequency.}
}
\value{
A bar chart of the distribution of consensus sRNA in the data.
}
\description{
Plot a bar chart of distribution of small RNA classes
(20-24nt sRNA) identified in the samples. Useful as a quality control step.
}
\examples{

data("sRNA_data_summary")
p1 <- plotConsensusFrequency(data = sRNA_data_summary, relative=TRUE )


}
