% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plots (logFC, FDR)}
\usage{
plotVolcano(data, labels = FALSE, top.molecules = 10, colour.scheme = NULL)
}
\arguments{
\item{data}{Dataframe; Must follow the structure created by
\code{mobileRNA::RNAimport()} and contain columns contributed by statistical
analysis using \code{mobileRNA::RNAanalysis()}}

\item{labels}{Add labels molecules with the most significant changes}

\item{top.molecules}{numeric; number of molecules to label}

\item{colour.scheme}{colour scale; recommend picking a
"Viridis colour scales from viridisLite". Default set to
ggplot2::scale_colour_viridis_d}
}
\description{
Volcano plots shows the relationship between fold change
and adjusted p-value (i.e. FDR). Each point represents an individual genes or
cluster. Pointes are coloured according to their significance level
thresholds (0.05, 0.01, and 0.001 or unchanged).
}
