% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{reduced_chr12_Eggplant.fa}
\alias{reduced_chr12_Eggplant.fa}
\title{reduced_chr12_Eggplant.fa}
\usage{
base::system.file("extdata","reduced_chr12_Eggplant.fa",package="mobileRNA")
}
\value{
FASTA file
}
\description{
Sample FASTA file
}
\details{
Reduced nucleotide sequence of chromosome 12 of
Eggplant (Solanum melongena) from Sol Genomics Network
(https://solgenomics.net/)
}
\examples{
 reduced_chr12_Eggplant <- base::system.file("extdata",
 "reduced_chr12_Eggplant.fa",package="mobileRNA")
}
\keyword{FASTA}
