% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{reduced_chr12_Eggplant.gff}
\alias{reduced_chr12_Eggplant.gff}
\title{reduced_chr12_Eggplant.gff}
\usage{
base::system.file("extdata","reduced_chr12_Eggplant.gff",package="mobileRNA")
}
\value{
GFF file
}
\description{
Sample GFF file
}
\details{
A small selection of genomic features from chromosome 12 of
Eggplant (Solanum melongena) from Sol Genomics Network
(https://solgenomics.net/)
}
\examples{
 reduced_chr12_Eggplant <- base::system.file("extdata",
 "reduced_chr12_Eggplant.gff",package="mobileRNA")
}
\keyword{GFF}
