% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{reduced_chr2_Tomato.fa}
\alias{reduced_chr2_Tomato.fa}
\title{reduced_chr2_Tomato.fa}
\usage{
base::system.file("extdata","reduced_chr2_Tomato.fa",package="mobileRNA")
}
\value{
FASTA file
}
\description{
Sample FASTA file
}
\details{
Reduced nucleotide sequence of chromosome 2 of Tomato
(Solanum lycopersicum) from Sol Genomics Network
(https://solgenomics.net/)
}
\examples{
 reduced_chr2_Tomato <- base::system.file("extdata",
 "reduced_chr2_Tomato.fa",package="mobileRNA")
}
\keyword{FASTA}
