% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{reduced_chr2_Tomato.gff}
\alias{reduced_chr2_Tomato.gff}
\title{reduced_chr2_Tomato.gff}
\usage{
base::system.file("extdata","reduced_chr2_Tomato.gff",package="mobileRNA")
}
\value{
GFF file
}
\description{
Sample GFF file
}
\details{
A small selection of genomic features from chromosome 2 of
Tomato (Solanum lycopersicum) from Sol Genomics Network
(https://solgenomics.net/)
}
\examples{
 reduced_chr2_Tomato <- base::system.file("extdata",
 "reduced_chr2_Tomato.gff",package="mobileRNA")
}
\keyword{GFF}
