% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceInfo.R
\name{referenceInfo}
\alias{referenceInfo}
\title{Extract chromosome sizes}
\usage{
referenceInfo(genome, out)
}
\arguments{
\item{genome}{Path to a genome (\code{.fa/.fasta})}

\item{out}{Path to output directory}
}
\value{
A plain text file listing the chromosomes and their respective sizes.
}
\description{
Invokes an OS command to calculate the chromosome sizes within
a given genome.
}
\examples{
\dontrun{
ref <- "/Users/user1/projectname/workplace/reference/merge/ref_merged.fa"
save <- "/Users/user1/projectname/workplace/reference/merge/"
chrInfo(genome = ref, out = save)
}

}
