% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAconsensus.R
\name{RNAconsensus}
\alias{RNAconsensus}
\title{Define the siRNA consensus for each dicer-derived cluster}
\usage{
RNAconsensus(data, conditions, tidy = TRUE)
}
\arguments{
\item{data}{data frame or GRanges object containing sample data. See
[sample_table()] to produce an organised GRanges object of sample data.}

\item{conditions}{Vector containing names of the heterografted conditions.
Or, samples you wish to draw the consensus class from in the analysis.}

\item{tidy}{Whether to remove clusters with an unknown or unclassified
consensus result. `tidy=TRUE` to remove unclassified cluster, or
\code{tidy=FALSE} to not remove background noise.
It is preferable to remove the excess noise in data.}
}
\value{
A data frame containing all existing columns in the input data object,
plus, an additional column labeled `sRNA_Consensus` stating the consensus
small RNA type/class between 20-24 nucleotides in length.
}
\description{
Using the data, the function creates a data frame with an
additional column stating the consensus siRNA class/type for each
dicer-derived cluster `sRNA_Consensus`.
}
\examples{

 data("sRNA_data")

# define consensus sRNA classes.
conditions <- c("TomEgg_1", "TomEgg_2", "TomEgg_3")

# Run function to define sRNA class for each cluster.
sRNA_data_summary <- RNAconsensus(data = sRNA_data,
                                     conditions = conditions,
                                    tidy=TRUE)

}
