% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAdistribution.R
\name{RNAdistribution}
\alias{RNAdistribution}
\title{Extract the distribution of RNA classes within each sample}
\usage{
RNAdistribution(
  data,
  style = c("bar", "line"),
  facet = TRUE,
  facet.arrange = 3,
  colour = NULL,
  together = TRUE
)
}
\arguments{
\item{data}{a dataframe, on which one of the following functions has already
been called: [RNAimport()],[RNAconsensus()].}

\item{style}{plotting option to choose the style of a line graph or bar chart
to represent your data. Where `style="line"` a line graph will be produced,
while `style=bar` produces a bar graph.}

\item{facet}{Logical; forms a matrix of panels defined by row and column faceting
variables. It plots the results for each sample as a bar-chart and contains
it within a single plot. The number of rows in the facet can be changed using
the augment ` facet.arrange` . Default ` facet = TRUE` , plot each
sample separately when ` facet = FALSE` .}

\item{facet.arrange}{numeric sent to the ` ncol`  augment in
[facet_wrap()] to define the number of columns.}

\item{colour}{bar plot fill colour. Default colour is "darkblue".}

\item{together}{Logical; forms a single line graph with multiple lines each
to represent the sample replicates. Default `together=TRUE`.}
}
\description{
` RNAdistribution`  extracts information from the RNA summary
data and calculates the total number of each RNA class identified within a
sample, for all samples. The results are plotted as either bar chart, with
an option to facet all plots into a single plot, or as a line graph.
}
\examples{
data('sRNA_data')
sample.distribution <- RNAdistribution(data = sRNA_data, style = "line")
}
