% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmean.R
\name{RNAmean}
\alias{RNAmean}
\title{Calculate mean RPM and mean Count across given samples}
\usage{
RNAmean(data, conditions = NULL)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAimport]{RNAimport()}}}

\item{conditions}{character; optional, if supplied represent sample names. If
not supplied, function will calculate means across all samples in dataframe.}
}
\value{
A data frame containing all existing columns in the input data object,
plus, an additional columns.

As default, the FPKM means will be stored in the column named \code{mean_RPM},
while the count means are stored in column names \code{mean_Count}.
}
\description{
Using the data, the function calculates the mean RPM and Counts
across all or specific samples.
}
\examples{
data("sRNA_data")
# across specific samples
selected_samples <- c("heterograft_1", "heterograft_2", "heterograft_3")

means <- RNAmean(data = sRNA_data, conditions = selected_samples)

# for all samples
means <- RNAmean(data = sRNA_data)

}
