% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmean.R
\name{RNAmeans}
\alias{RNAmeans}
\title{Calculate FPKM across samples}
\usage{
RNAmeans(data, conditions = NULL)
}
\arguments{
\item{data}{data frame, see [RNAlocate::RNAimport()] to produce an organised
object of sample data.}

\item{conditions}{Character vector; represent sample names. If not supplied,
function will calculate means across all samples in dataframe.}
}
\value{
A data frame containing all existing columns in the input data object,
plus, an additional columns.

As default, the FPKM means will be stored in the column named `mean_RPM`,
while the count means are stored in column names `mean_Count`.
}
\description{
Using the data, the function calculates the mean RPM and Counts
across all or specific samples.
}
\examples{
data("sRNA_data")
# across specific samples
selected_samples <- c("TomEgg_1", "TomEgg_2", "TomEgg_3")
means <- RNAmean(data = sRNA_data, conditions = selected_samples)
# for all samples
means <- RNAmean(data = sRNA_data)
}
