% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmergeAnnotations.R
\name{RNAmergeAnnotations}
\alias{RNAmergeAnnotations}
\title{Merge two GFF genome annotation files}
\usage{
RNAmergeAnnotations(
  annotationA,
  annotationB,
  out_dir,
  abbreviationAnnoA = "A",
  abbreviationAnnoB = "B",
  exportFormat = "GFF"
)
}
\arguments{
\item{annotationA}{path; directory path to genome annotation assembly file in
GFF format.}

\item{annotationB}{path; directory path to genome annotation assembly file in
GFF format.}

\item{out_dir}{path; a character string or a \code{base::connections()} open
for writing. Including file output name, and file extension of \code{.fa}.}

\item{abbreviationAnnoA}{character; string to represent prefix added to
existing chromosome names in \code{annotationA}. Default set as "A", which is
separated from existing chromosome names by an underscore (_).}

\item{abbreviationAnnoB}{character; string to represent prefix added to
existing chromosome names in \code{annotationB}. Default set as "B", which is
separated from existing chromosome names by an underscore (_).}

\item{exportFormat}{The format of the annotation output. If missing or the
format does not match the file type, an error will occur. Default set to "GFF"
format.}
}
\value{
The function output the individually altered annotation files, plus, a merged
annotation file to the global environment and the given directory.
The altered annotation files contain either default or custom alterations to
the chromosome names and are written to the same directory as the merged file,
with the given names addition of \code{annotationA_altered.gff} and
\code{annotationB_altered.gff}, respectively.
}
\description{
Merges two genomes annotation files (\verb{GFF format}) into one
single GFF file saved to desired directory. Adds prefix to each chromosome
name in the genome annotations to ensure each assembly is distinguishable
within the merged file, and saved the individually altered references to
local directory.
}
\details{
The functions primary goal is to merge two GFF files, however, when
merging genomic files it is critical that the two genomes are distinguishable
by the chromosome names. As a default setting, the function extracts the
chromosome names for a given GFF file and adds an ID before each chromosome
name and removes any periods.

As default, the function as a prefix to chromosomes which contains a either
"A" or "B" and "\emph{". It will rename the chromosome names in \code{annotationA}
to "A}". For example, A_0, A_1, A_2 etc. To set a custom chromosome name for
\code{annotationA} alter the argument \code{abbreviationAnnoA}. While, for
\code{annotationB} as default the chromosome names will be named "B_",
for example, B_0, B_1, B_2 etc. To set a custom chromosome name for
\code{annotationB} alter the argument \code{abbreviationAnnoB}. Please note that
the underscore is added automatically, hence, when setting a custom prefix
just include character values.

The merged genome is saved to the specified output directory, and requires
the user to set the name with a \code{gff} extension. The user must load the
annotations into the R Global Environment, hence, the alterations made by
this function will alter the variable. The altered annotations will also
be saved to the same directory as supplied for the output file, with a fixed
output name of \code{annotationA_altered.gff} and \code{annotationB_altered.gff},
respectively.

\emph{IMPORTANT:} The genome reference and annotation of a species
must have chromosomes with matching names. It is critical that if you used
the \code{\link[=RNAmergeGenomes]{RNAmergeGenomes()}} function to to create a merged reference
genome,that you treat the input annotations in the same way.
}
\examples{
# Load BiocFileCache package
library(BiocFileCache)

# generate cache in mobileRNA package
package_directory <- system.file(package = "mobileRNA")

# generate custom folder in package directory
cache_directory <- file.path(package_directory, "cache")

# Initialize a BiocFileCache in the cache directory
cache <-  BiocFileCache(cache_directory, ask = FALSE)

# Generate URL to GFF files
url_remote <- "https://github.com/KJeynesCupper/assemblies/raw/main/"

anno1 <- paste0(url_remote,"chr12_Eggplant_V4.1_function_IPR_final.gff")
anno2 <- paste0(url_remote, "chr2_ITAG4.0_gene_models.gff")

# Add GFF files to cache:
add_anno1 <- bfcadd(cache,"chr12_Eggplant_V4.1_function_IPR_final", fpath=anno1)
rid1 <- names(add_anno1)

add_anno2 <- bfcadd(cache,"chr2_ITAG4.0_gene_models",fpath=anno2)
rid2 <- names(add_anno2)

# rid1 and rid2 object contain the path location to the GFF files 
# edit and merge
merged_anno <- RNAmergeAnnotations(annotationA = cache[[rid1]],
                                   annotationB = cache[[rid2]],
             out_dir = tempfile("merged_annotation",tmpdir = cache_directory, 
             fileext = ".gff3"))

}
