% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmobile.R
\name{RNAmobile}
\alias{RNAmobile}
\title{Identify putative RNA molecules produced by the non-tissue sample genome}
\usage{
RNAmobile(
  input = c("sRNA", "mRNA"),
  data,
  controls,
  genome.ID,
  task = NULL,
  statistical = FALSE,
  alpha = 0.1,
  threshold = NULL,
  baymobil_integration = FALSE,
  baymobil_filepath = NULL,
  annotation_filepath = NULL,
  gene_id_field = "ID",
  log10BF_threshold = 1,
  cap = c(-2, 10)
)
}
\arguments{
\item{input}{character; must be either "sRNA" or "mRNA" to represent the type
of data.}

\item{data}{data.frame; generated through the \pkg{mobileRNA} method.}

\item{controls}{character vector; containing names of control samples.}

\item{genome.ID}{character; string or chromosome identifier related to the
chromosomes in a given genome. A distinguishing feature of the genome of
interest or non-interest in the chromosome name (\code{chr} column).}

\item{task}{character; string to set the method to keep or remove the
chromosomes containing the identifying string. To keep the chromosomes with
the ID, set task=keep. To remove, set \code{task="remove"}. As default, task is
set to \code{keep}.}

\item{statistical}{If TRUE, will undertake statistical filtering based on the
a p-value or adjusted p-value threshold stated by \code{alpha}.Default
set at FALSE. Requires presence of columns containing statistical data.
In order to filter by the adjusted p-value, a column named \code{padjusted} must
be present. See \code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}} to calculate
statistical values.}

\item{alpha}{numeric; adjusted p-value cutoff as the target FDR for
independent filtering. Default is 0.1. Only mobile molecules with adjusted
p-values equal or lower than specified are returned.}

\item{threshold}{numeric; set a threshold level. For sRNA analysis, this
represents filtering by the minimum number of replicates that defined the
consensus dicercall which is stored in the \code{DicerCounts} column. While,
for mRNA analysis this represents the number of replicates which contained
reads for the mRNA molecule which is stored in the \code{SampleCounts} column.}

\item{baymobil_integration}{logical; state whether to supply results from
baymobil (Default = FALSE)}

\item{baymobil_filepath}{path; directory path to a CSV file produced as
output from baymobil.Only required for baymobil integration.}

\item{annotation_filepath}{path; directory path to the merged annotation file
(GFF).Only required for baymobil integration.}

\item{gene_id_field}{character; column in annotation file supplying gene
names or ID (Default ="ID"). Only required for baymobil integration.}

\item{log10BF_threshold}{numeric; summed log10BF value to use as threshold
for baymobil integration. Only required for baymobil integration.}

\item{cap}{vector; vector containing lowest and highest numeric values to
cap log10BF values. (Default = c(-2, 10)) Only required for baymobil
integration.}
}
\value{
A data frame containing candidate mobile sRNAs or mRNAs, which could
have been further filtered based on statistical significance and the ability
to by-pass the thresholds which determine the number of replicates that
defined the consensus dicercall (sRNA) or contributed to reads counts (mRNA).

If baymobil results are integrated for mRNA mobility calculation, then
additional columns are added from the baymobil output. Only mRNA transcripts
with supported statistical evidence for mobility are added. In some cases,
this will include genes that are not detected as mobile by mobileRNA. The
column "baymobil_mobility_status" is always added, and states whether the
mRNA transcript has evidence for mobility from baymobil or whether it has no
evidence. When this is used the statistical and threshold
filtering is not applied.
}
\description{
A function to identify the putative sRNA or mRNA  molecules
produced by the non-tissue sample genome. Includes putative RNA mobilome or
RNAs not expected to be found within the tissue of origin.
}
\details{
The function identifies candidate sRNAs or mRNAs produced by a specific
genome/genotype. It does so by either keeping or removing those mapped to a
given genome. To do so, it requires a common pre-fix across chromosomes of
the given genome. See\code{\link[=RNAmergeGenomes]{RNAmergeGenomes()}} for more information.
In addition, it removes RNAs which were likely to be falsely mapped. These
are those which were mapped to the non-tissue genotype in the control
samples.

\strong{For sRNAseq:}
A greater confidence in the sRNA candidates can be achieved by setting
a threshold that considers the number of replicates which contributed to
defining the consensus dicercall (ie. consensus sRNA classification). This
parameter filters based on the \code{DicerCounts} column introduced by the
\code{\link[=RNAdicercall]{RNAdicercall()}} function.

\strong{For mRNAseq:}
A greater confidence in the mRNA candidates can be achieved by setting
a threshold that considers the number of replicates which contained reads
for the mRNA molecule. This parameter filters based on the \code{SampleCounts}
column introduced by the \code{\link[=RNAimport]{RNAimport()}} function.

For intraspecfic grafts (eg. A.thaliana ecotypes), we have demonstrated it to
be favourable to utilize the \code{baymobil} algorithm alongside mobileRNA.
This approach utilities exact Bayesian inference based on SNP variant
information. Please run the \code{baymobil} outside of mobileRNA, and here, you
can supply the output of \code{baymobil} to add more information to the mobileRNA
results. Please refer to our latest publication for a comprehensive breakdown.

To determine mobility, the log10 Bayes factors (logBF) is summed across SNPs
for each transcript, neglecting SNPs where Nh1 is zero and limiting the logBF
value to -2 and capping at 10 (default). Transcripts with a summed logBF
greater than or equal to the threshold value (log10BF_threshold,
default = 1) have statistical support for mobility.

\strong{Statistical Analysis}
The function also allows for filtering using statistical inference generated
from the differential analysis of the total data set using the function
\code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}}. When \code{statistical=TRUE}, the feature
is enabled and selects molecules that meet the adjusted p-value
cutoff defined by \code{alpha}.
}
\examples{
data("sRNA_data")


# vector of control names
controls <- c("selfgraft_1", "selfgraft_2" , "selfgraft_3")

# Locate potentially mobile sRNA clusters associated to tomato, no
# statistical analysis
mobile_df1 <- RNAmobile(input = "sRNA", data =  sRNA_data,
controls = controls, genome.ID = "B_", task = "keep", statistical = FALSE)



}
\references{
baymobil \url{https://doi.org/10.21203/rs.3.rs-2520491/v1}
}
