% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsignificant.R
\name{RNAsignificant}
\alias{RNAsignificant}
\title{Extract statistically significant sRNA clusters within a population.}
\usage{
RNAsignificant(
  data,
  statistical = FALSE,
  padj = 0.05,
  p.value = NULL,
  chimeric = FALSE,
  controls = NULL,
  genome.ID = NULL
)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAimport]{RNAimport()}}}

\item{statistical}{logical; undertake statistical filtering based on the
a p-value or adjusted p-value threshold stated by \code{padj} & \code{p.value} when
\code{statistical=TRUE}. Default setting \code{statistical=FALSE}. Requires presence of
columns containing statistical data,see function
\code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}} to calculate statistical values.}

\item{padj}{numeric; user defined numeric value to represent the adjusted
p-value threshold to define statistic significance. Defaults setting
\code{padj=0.05}. Returns sRNA clusters with an adjusted p-values equal or lower
than the threshold value.}

\item{p.value}{numeric; user defined numeric value to represent the p-value
threshold to define statistic significance. There is no default value, set
this instead of using an adjusted p-value to filter sRNA cluster.}

\item{chimeric}{logical; state whether system is chimeric: contains multiple
genomes/genotypes.}

\item{controls}{character; vector of control condition sample names.}

\item{genome.ID}{character; chromosome identifier of foreign genome in
chimeric system}
}
\value{
A refined version of the working dataframe supplied to the function.
The function selects sRNA clusters which meet the statistical threshold,
given the statistical analysis has been undertaken using the
\code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}} function.
}
\description{
sRNA clusters which are more abundant, show a statistically
significant logfc.
}
\details{
Filters the data based on statistical significance
which has been calculated by the \code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}}
function. This optional features enables the user to select sRNA clusters
which meet a specific p-value or adjusted p-values threshold.

Requires either or both columns: \code{pvalue}, \code{padjusted} to undertake.

When working with a chimeric system, for example interspecific grafting,
mapping errors can easily be recognised and eliminated. Here, these can be
eliminated by supplying some extra parameter information. State
\code{chimeric=TRUE} and supply the chromosome identifier of the foreign genome
(ie. not the tissue sample genotype, but the genotype from which any
potential mobile molecules could be traveling from) to the \code{genome.ID}
parameter & the control condition samples names to the \code{controls} parameter.
}
\examples{

data("sRNA_data_consensus")

## sample conditions in order within dataframe
groups <- c("Heterograft", "Heterograft", "Heterograft",
            "Selfgraft", "Selfgraft", "Selfgraft")


## Differential analysis using the DEseq2 method 
sRNA_DESeq2 <- RNAdifferentialAnalysis(data = sRNA_data_consensus,
                          group = groups,
                          method = "DESeq2")
                          
 ## Select significant based on padjusted                        
significant_sRNAs <- RNAsignificant(sRNA_DESeq2, chimeric = TRUE, 
                                    genome.ID = "SL", 
                                    controls = c("selfgraft_1","selfgraft_2", 
                                    "selfgraft_3"))
                                 
                                 

}
