% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsubset.R
\name{RNAsubset}
\alias{RNAsubset}
\title{Subset data to select a specific small RNA population}
\usage{
RNAsubset(data, type, sig = FALSE, ...)
}
\arguments{
\item{data}{A numerical data-frame containing the sample data and consensus
sRNA class/type (see [RNAconsensus()].}

\item{type}{A number to represent the type of small RNA population to subset
for.}

\item{sig}{Parameter to filter and select significant sRNA. If
\code{sig=TRUE}, data will be filtered based on p-adjusted < 0.05
significance threshold.}

\item{...}{Related to number in the `type` argument
This can be a value from 20-24. To select, 24-nt sRNA, state 24.
Multiple values can be inputted, for instance both 21 and 22 can be
stated to select both.}
}
\description{
Creates a data-frame containing the desired sRNA class(s).
See [RNAconsensus()] for information on defining the sRNA type
for each cluster. The function allows the choice to filtered the data by
statistical significance. Set \code{sig=TRUE} to filtered by significance
(p-adjusted).
}
\examples{
data("sRNA_data")

##  define consensus sRNA classes.
samples <- c("TomEgg_1", "TomEgg_2", "TomEgg_3")

# Define consensus
sRNA_data_summary <- RNAconsensus(data = sRNA_data,
                                     conditions = samples)

# Subset data for  24-nt sRNAs
sRNA_24 <- RNAsubset(sRNA_data_summary, type = 24)


# Subset data for 24 21/22-nt sRNAs
sRNA_2122 <- RNAsubset(sRNA_data_summary, type = c(21, 22))

# You can subset by any combination of classes. For example, a dataset
# of 23-nt & 24-nt sRNAs or just 20-nt sRNAs.

}
