% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{selfgraft_demo_2.fq}
\alias{selfgraft_demo_2.fq}
\title{selfgraft_demo_2.fq}
\usage{
base::system.file("extdata/sRNAseq","selfgraft_demo_2.fq",package="mobileRNA")
}
\value{
FASTQ file
}
\description{
Demo FASTQ file
}
\details{
Demo sRNAseq FASTQ file for a self-graft control Eggplant
(Solanum melongena) sample. This file has been majorly reduced and does not
represent true data. For pre-processing purposes.
}
\examples{
 selfgraft_demo_2 <- base::system.file("extdata/sRNAseq","selfgraft_demo_2.fq",
 package="mobileRNA")
}
\keyword{FASTQ}
