% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{data}
\name{sRNA_data_mobile}
\alias{sRNA_data_mobile}
\title{sRNA_data_mobile: simulated data for biological replicates}
\usage{
data(sRNA_data_mobile)
}
\value{
Dataframe in global environment
}
\description{
Simulated sRNAseq dataset - potentially mobile RNAs
}
\details{
Simulates data is taken from eggplant and tomato sRNAseq samples and
created to simulate to movement of sRNA molecules from an Tomato rootstock to
an Eggplant Scion. Three Eggplant replicates were spiked with the same 150
tomato sRNA clusters, and named "heterograft_" 1 to 3. The analysis compares
these heterografts to three Eggplant self-grafts which are the original
un-spiked Eggplant replicates, called "selfgraft_" 1 to 3.

This data is a further development of the \code{sRNA_data} data set, where the
candidate mobile sRNAs have been selected.
}
\examples{
 data("sRNA_data_mobile")
}
\keyword{dataset}
