% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRNAmapper.R
\name{sRNAmapper}
\alias{sRNAmapper}
\title{Map & cluster sRNAseq reads using \code{ShortStack}}
\usage{
sRNAmapper(
  input_files_dir,
  output_dir,
  genomefile,
  threads = 6,
  mmap = "u",
  dicermin = 20,
  dicermax = 25,
  mincov = 2,
  pad = 75,
  tidy = TRUE
)
}
\arguments{
\item{input_files_dir}{path; directory containing only the FASTQ sRNAseq
samples for analysis. Note that all samples in this directory will be used by
this function.}

\item{output_dir}{path; directory to store output.}

\item{genomefile}{path; path to a FASTA file.}

\item{threads}{numeric; set the number of threads to use where more threads
means a faster completion time. Default is 6.}

\item{mmap}{character; define how to handle multi-mapped reads. Choose from
"u", "f" or "r". Where "u" means only uniquely-aligned reads are used as
weights for placement of multi-mapped reads. Where "f" means fractional
weighting scheme for placement of multi-mapped reads and "r" mean
multi-mapped read placement is random. Default is "u".}

\item{dicermin}{integer; the minimum size in nucleotides of a valid small
RNA. This option sets the bounds to discriminate dicer-derived small RNA loci
from other loci. Default is 20.}

\item{dicermax}{integer; the maximum size in nucleotides of a valid small
RNA. This option sets the bounds to discriminate dicer-derived small RNA loci
from other loci.  Default is 25.}

\item{mincov}{numeric; minimum alignment depth, in units of reads per
million, required to nucleate a small RNA cluster during de novo cluster
search. Must be an floating point number > 0. Default is 2.}

\item{pad}{interger; initial peaks are merged if they are this distance or
less from each other. Must >= 1, default is 75.}

\item{tidy}{logical; removes unnecessary extra output files when set to TRUE.}
}
\description{
Map sRNAseq FASTQ samples to a given genome assembly using
ShortStack.
}
\details{
This function undertakes our recommended workflow for mapping sRNAseq reads
to a given genome for the analysis of small RNAs in a biological system.

The function invokes a number of OS commands, and is dependent on the
installation of \code{ShortStack} and it's dependencies within the same
environment. Please refer to \link{} for more information. Please not that
\code{ShortStack} is only compartible with Linux and Mac operating systems.

It uses \code{ShortStack} to map and cluster sRNAs for each sample
within a directory. Within the output directory (\code{output_dir}), the function
generates two new directories called  "1_map" and "2_map", plus, a plain
text file called "locifile.txt". The function undertakes an initial mapping
step in order to generate a list of de novo sRNA-producing loci. The results
for each sample are saved to "1_map", producing one folder for each sample.
The de novo sRNA-producing loci identified in each sample are merged;
generating the "locifile.txt" file. The second mapping step utilises the
"locifile.txt" file to optimise the analysis. The results of mapping step two
are saved to the "2_map" directory, where one folder for each sample is
generated and stores the results.

The function generates a number of extra files for each sample and are not
required for the downstream analysis. Hence, as default these files are
deleted. This is determined by the \code{tidy} argument.
}
\examples{
\dontrun{

options(shortstack_path = "[path to shortstack]")

samples <- file.path(system.file("extdata",package="mobileRNA"))

output_location <- tempdir()

sRNAmapper(input_files_dir = samples, 
output_dir = output_location, 
genomefile = output_assembly_file)
}




}
