% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFiles.R
\name{mergeFiles}
\alias{mergeFiles}
\title{Merge two reference genomes into one}
\usage{
mergeFiles(files, file1 = NULL, file2 = NULL, out)
}
\arguments{
\item{files}{Path to the directory containing genome both reference files
(The directory must only contain these files).}

\item{file1}{Path to the directory of the first genome reference file.}

\item{file2}{Path to the directory of the second genome reference file.}

\item{out}{Path desired output directory.}
}
\value{
A file named \code{ref_merged.fa} representing the merged
genome reference.
}
\description{
Merges two reference genomes (`.fa/.fasta)`. into one single
reference.
}
\details{
Prior to merging it is import to check the chromosome names for each
genome. This is to ensure they can be easily distinguished from each other
after merging. To extract the chromosomal information, look at
[RNAlocate::referenceInfo()] and to modify the chromosome names, look at
[RNAlocate::modifyReference()].
}
\examples{
\dontrun{
mergeFiles(file1 = "./workplace/reference/ref1.fa.gz",
            file2 = "./workplace/reference/ref2.fa.gz"
            out = "./workplace/reference/merge/ref_merged.fa")
# Or,
mergeFiles(files = "./workplace/reference/*",
            out = "./workplace/reference/merge/ref_merged.fa")
}


}
