% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileRNA.R
\docType{package}
\name{mobileRNA}
\alias{mobileRNA}
\title{mobileRNA: Explore RNA mobilome & population-scale changes}
\description{
Uses small RNA or messenger RNA sequencing data in two conditions and
identifies changes in the RNA population. mobileRNA was primarily designed
for the identification of a putative RNA mobilome in a chimeric system. For
example, in plant graft systems. As input, mobileRNA takes sRNAseq or mRNAseq
fastq files. Output consists of a dataframe with putative differences
between two conditions along with a number of plots.
}
\details{
View the vignette here:
\href{https://github.com/KJeynesCupper/mobileRNA/tree/devel/docs/articles/mobileRNA.html}{mobileRNA}

The most important functions in the \pkg{mobileRNA} are:
\describe{
\item{\code{\link{RNAmergeGenomes}}}{Merge two genome assembly files (FASTA)}
\item{\code{\link{RNAmergeAnnotations}}}{Merge two genome annotation files
(GFF)}
\item{\code{\link{mapRNA}}}{Pre-processing of sRNAseq and mRNAseq
(alignment, raw count, cluster analysis)}
\item{\code{\link{RNAimport}}}{Reads the pre-processing report files in
to a dataframe for all conditions}
\item{\code{\link{RNAdicercall}}}{Calculates the consensus sRNA
dicercall class}
\item{\code{\link{RNAsubset}}}{Subsets the working data set to select a
defined sRNA class}
\item{\code{\link{RNAdifferentialAnalysis}}}{Undertakes differential
analysis using either the edgeR or DESeq2 method for data structure}
\item{\code{\link{RNAmobile}}}{Identifies mobile candidates in chimeric
systems}
\item{\code{\link{RNApopulation}}}{Selects unique RNAs within the defined
sample replicates}
\item{\code{\link{RNAsummary}}}{Summarise the differential abundance of RNAs}
\item{\code{\link{RNAreorder}}}{Reorder the data frame for differential
analysis, ensuring control verse treatment comparison}
\item{\code{\link{RNAsequences}}}{Defines the consensus RNA sequence for a
sRNA cluster}
\item{\code{\link{RNAattributes}}}{Overlaps the sRNA clusters with a GFF
annotation file and adds overlapping features to the sRNA cluster
information}
\item{\code{\link{RNAdistribution}}}{Plot the distribution of sRNA
lengths/classes}
\item{\code{\link{plotHeatmap}}}{Plots a heatmap of log2-transformed
normalised RPM/FPKM values}
\item{\code{\link{plotSampleDistance}}}{Plots a sample distance heatmap for
quality control}
\item{\code{\link{plotSamplePCA}}}{Plots a PCA plot, customise ratio,
colours and shapes}
\item{\code{\link{RNAfeatures}}}{Overlaps the sRNA clusters with a GFF
annotation file and calculates the percentage of clusters which overlap with
specific genomic features}
\item{\code{\link{RNAdf2se}}}{Convert \code{mobileRNA} dataframe to
a SummarizedExperiment object }
}
}
\seealso{
See \code{vignette("mobileRNA", package = "mobieRNA")} for an
overview of the package.
}
\author{
Katie Jeynes-Cupper \email{kej031@student.bham.ac.uk},
Marco Catoni \email{m.catoni@bham.ac.uk}
Maintainer: Katie Jeynes-Cupper \email{kej031@student.bham.ac.uk}
}
