% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap using hierarchical clustering}
\usage{
plotHeatmap(data, colours = NULL, dendogram = TRUE, margins = NULL)
}
\arguments{
\item{data}{Data frame containing FPKM values for each samples on
each sRNA dicer-derived cluster of interest.}

\item{colours}{Colours to display and represent the heatmap.
Defaults to [grDevices::heat.colors()] (heat.colors(100)).}

\item{dendogram}{Logical; indicating whether to include the dendrogram, and
retain clustering. Default, \code{dendogram = TRUE} to include.}

\item{margins}{Numeric vector; Length of 2, to state width of the heatmap
column names section and row names section, respectively. If null, default
`margins` is set as c(10, 10).}
}
\description{
Plots a heatmap with hierarchical clustering via an rlog
transformation of FPKM data and euclidean statistics.
}
\details{
The function create a heatmap based on the hierarchical clustering
of FPKM values using euclidean statistics.
}
\examples{

## DESeq2 example: mobile 24-nt & 21/22-nt sRNA

 # plot heatmap of likely mobile 24-nt sRNA
 data("sRNA_24_mobile_DESeq2")
 p1 <-  plotHeatmap(sRNA_24_mobile_DESeq2)

 # plot heatmap of likely mobile 24-nt sRNA
 data("sRNA_2122_mobile_DESeq2")
 p2 <-  plotHeatmap(sRNA_2122_mobile_DESeq2)



## edgeR example: mobile 24-nt & 21/22-nt sRNA

 # plot heatmap of likely mobile 24-nt sRNA
 data("sRNA_24_mobile_edgeR")
 p3 <-  plotHeatmap(sRNA_24_mobile_edgeR)

 # plot heatmap of likely mobile 24-nt sRNA
 data("sRNA_2122_mobile_edgeR")
 p4 <-  plotHeatmap(sRNA_2122_mobile_edgeR)

}
