% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap of log-transformed RPM data}
\usage{
plotHeatmap(
  data,
  value = "RPM",
  pseudocount = 1e-06,
  colours = grDevices::hsv(1, 1, seq(0, 1, length.out = 12)),
  cluster = TRUE,
  scale = "none",
  clustering_method = "complete",
  row.names = TRUE,
  border.color = NA
)
}
\arguments{
\item{data}{Dataframe; Must follow the structure created by
\code{mobileRNA::RNAimport()}.}

\item{value}{the values to plot, either \code{FPKM} or \code{RPM}}

\item{pseudocount}{numeric; pseudocount, default is \code{1e-6}}

\item{colours}{vector of colors. Default is
\code{grDevices::hsv(1,1,seq(0,1,length.out = 12))}}

\item{cluster}{logical; include hierarchical clustering when default
\code{cluster= TRUE}}

\item{scale}{character; indicating whether the values should be centered &
scaled in either the row direction or the column direction, or none.
Respective options are "row", "column" & "none". Default, \code{scale="none"}.}

\item{clustering_method}{Character; clustering method used. Accepts the same
values as hclust. Default \code{clustering_method= "complete"}}

\item{row.names}{logical; indicated whether to include cluster names as
rownames. Default \code{row.names=TRUE}}

\item{border.color}{border colour, default is no border, NA.}
}
\value{
Produces a list of objects, including the plot.
}
\description{
Undertakes RPM/FPKM normalisation using a pseudocount and
transforms the data using log-scale, to enable visualization of the
differences and patterns in expression across samples using a heatmap.
}
\details{
Undertakes FPKM/RPM normalisation using a pseudocount and then
transforms the normalised-RPM data using log-scale. The
\code{mobileRNA::RNAimport()} function will  include the RPM data
columns for sRNAseq data processed by ShortStack, while with mRNAseq data
the function calculates FPKM.
The data is then plotted as a heatmap, utilising the \code{pheatmaps} package.

This function expects to receive a dataframe containing FPKM/RPM data from
sRNA-seq studies. This function employs the use of a pseudocount during
normalisation as the function is expected to be used when identifying mobile
sRNAs in a chimeric system. In such system, it is expected that control
replicates will contain zero values for the candidate mobile sRNA clusters.
}
\examples{

data("sRNA_data_mobile")

# plot heatmap of potential mobile sRNAs
 p1 <-  plotHeatmap(sRNA_data_mobile)




}
