% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap of log-transformed normalization data}
\usage{
plotHeatmap(
  data,
  value = "RPM",
  pseudocount = 1e-06,
  colours = (grDevices::colorRampPalette(RColorBrewer::brewer.pal(9, "Blues")))(100),
  cluster = TRUE,
  scale = "none",
  clustering_method = "complete",
  row.names = TRUE,
  border.color = NA
)
}
\arguments{
\item{data}{data.frame; originally generated by \code{\link[=RNAimport]{RNAimport()}}}

\item{value}{character; either \code{FPKM} or \code{RPM} - represents values to plot}

\item{pseudocount}{numeric; pseudocount, default is \code{1e-6}}

\item{colours}{character; colors. Default is
\code{grDevices::colorRampPalette(RColorBrewer::brewer.pal(9, "Blues"))(100)}}

\item{cluster}{logical; include hierarchical clustering when default
\code{cluster= TRUE}}

\item{scale}{character; indicating whether the values should be centered &
scaled in either the row direction or the column direction, or none.
Respective options are "row", "column" & "none". Default, \code{scale="none"}.}

\item{clustering_method}{character; clustering method used. Accepts the same
values as hclust. Default \code{clustering_method= "complete"}}

\item{row.names}{logical; indicated whether to include cluster names as
rownames. Default \code{row.names=TRUE}}

\item{border.color}{border colour, default is no border, NA.}
}
\value{
Produces a list of objects, including the plot.
}
\description{
Undertakes RPM/FPKM normalisation using a pseudocount and
transforms the data using log-scale, to enable visualization of the
differences and patterns in expression across samples using a heatmap.
}
\details{
Undertakes FPKM/RPM normalisation using a pseudocount and then
transforms the normalised-RPM data using log-scale.

This function expects to receive a dataframe containing FPKM/RPM data from
sRNA-seq studies. This function employs the use of a pseudocount during
normalisation as the function is expected to be used when identifying mobile
sRNAs in a chimeric system. In such system, it is expected that control
replicates will contain zero values for the candidate mobile sRNA clusters.
}
\examples{

data("sRNA_data_mobile")

# plot heatmap of potential mobile sRNAs
 p1 <-  plotHeatmap(sRNA_data_mobile, row.names = FALSE)




}
