% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleDistance.R
\name{plotSampleDistance}
\alias{plotSampleDistance}
\title{Heatmap to access sample distance as a quality control step}
\usage{
plotSampleDistance(data)
}
\arguments{
\item{data}{data.frame object containing raw count data}
}
\value{
A blue scale heatmap illustrating the sample distance
}
\description{
A function to draw a simple hierarchical clustered heatmap to observe
sample distance.
}
\examples{

data("sRNA_24")
plotSampleDistance(sRNA_24)

}
