% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAattributes.R
\name{RNAattributes}
\alias{RNAattributes}
\title{Identify specific attributes associated to a sRNA cluster}
\usage{
RNAattributes(data, annotation)
}
\arguments{
\item{data}{data frame; containing rows of potential dicer-derived clusters
including columns which supply the genomic coordinates, where `chr` supplies the
chromosome number, `start` and `end` which supply the coordinates.}

\item{annotation}{A path, URL, connection or GFFFile object. A genome
reference annotation file (.gff/.gff1/.gff2/.gff3).}
}
\description{
Based on genomic coordinates, assign sRNA clusters with an annotation.
}
\details{
The function merges the annotation (.gff/.gff3) information with
the sRNA data set based on the chromosome, start and end coordinates.
}
