% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAdicercall.R
\name{RNAdicercall}
\alias{RNAdicercall}
\title{Define the consensus dicercall for each sRNA cluster}
\usage{
RNAdicercall(
  data,
  conditions = NULL,
  ties.method = NULL,
  tidy = FALSE,
  chimeric = FALSE,
  controls = NULL,
  genome.ID = NULL
)
}
\arguments{
\item{data}{data.frame; originally generated by \code{\link[=RNAimport]{RNAimport()}}.}

\item{conditions}{character; vector containing sample replicate names. When
supplied, the data from the named replicates will be the only ones used to
calculate the dicercall consensus for each sRNA cluster. Each
string should represent a sample name present in the dataframe supplied to
the \code{data} argument.}

\item{ties.method}{character; string specifying how ties are handled, choose
either "exclude" or "random". When using \code{random}, if there is a tie one of
the classes will be choose at random. While, when using \code{exclude} if there is
a tie the class is set to undefined, however, if there is a tie between a
undefined and a known class, the known class takes president (eg 3x24-nt and
3xN-nt, then it will be classed as 24nt). Default setting
\code{ties.method="exclude"}.}

\item{tidy}{logical; tidy-up data by removing sRNA clusters with an unknown
or unclassified result. Default setting \code{tidy=TRUE}, removes excess
background noise.}

\item{chimeric}{logical; state whether system is chimeric: contains multiple
genomes/genotypes.}

\item{controls}{character; vector of control condition sample names.}

\item{genome.ID}{character; chromosome identifier of genome where mobile
molecules are produced in the chimeric system}
}
\value{
A data frame containing all existing columns in the input data object,
plus, two additional columns of data:

The first column, \code{DicerCounts} contains the number of replicates which
contributed to defining the dicer-derived sRNA class. This can be
utilised within the \code{RNAmobile} function as a threshold parameter.
While, utilising \code{exclude} ties methods, the \code{DicerCounts} value for a tie
will be 0 as no consensus classification was concluded. While, if the majority
of replicates contributed unclassified for a cluster, the \code{DicerCounts} value
represents the number of replicates which which contributed to the
unclassified consensus in \code{DicerConsensus}.

The second, labeled \code{DicerConsensus} states the consensus sRNA class between
nucleotides in length or "N" if unclassified.
}
\description{
The dicercall represent the length in nucleotides of the most
abundant sRNA sequence within a cluster. The function calculates the
consensus dicercall classification based on the predictions for the samples
in the analysis.
}
\details{
For each sample, the mapping step will have predicted the sRNA dicercall for
each cluster which is stored in the sample columns with names containing with
"DicerCall_".  The dicercall value represent the length in nucleotides of
the most abundant sRNA within the cluster. For some clusters, there is no
particular sRNA which is more abundant than another, hence, it is stated as
"NA" or "N", which is refereed to as being unclassified. Depending on the
organism the dicercall values will have varying ranges.

The \code{RNAdicercall()} function is used to calculate the consensus dicercall
for each sRNA cluster. This is based of the classification predicted for the
cluster by each sample within the analysis, described above. There are
several parameters which will alter the output, including the handling of
ties and the method to draw the consensus from.

When ties.method = "random", as per default, ties are broken at random.
In this case, the determination of a tie assumes that the entries are
probabilities: there is a relative tolerance of 1e-5, relative to the
largest (in magnitude, omitting infinity) entry in the row.

When ties.method = "exclude", ties between sRNA classification are ruled as
unclassified ("N"). However, when there is a tie between the choice of a
class or unclassified result the \code{exclude} option will always select the
class choice over the unclassified result.

If users are working with a chimeric system, utilise the \code{chimeric=TRUE}
parameter and state \code{genome.ID} and \code{controls} parameter variables. This will
remove any potential mapping errors which could introduce false interpretation.

To remove excess data noise, \code{tidy=TRUE} can be used to removed unclassified
("N") sRNA clusters, resulting in a reduced dataset size.
}
\examples{
 # load data 
 data("sRNA_data")

# define consensus sRNA classes.
conditions <- c("heterograft_1", "heterograft_2", "heterograft_3")

# Run function to define sRNA class for each cluster.
sRNA_data_dicercall <- RNAdicercall(data = sRNA_data,
                                  conditions = conditions,
                                  tidy=TRUE)

}
