% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAdifferentialAnalysis.R
\name{RNAdifferentialAnalysis}
\alias{RNAdifferentialAnalysis}
\title{Differential Analysis of sRNA clusters or mRNA using \code{DESeq2} or \code{edgeR}}
\usage{
RNAdifferentialAnalysis(
  data,
  group,
  method = c("edgeR", "DESeq2"),
  dispersionValue = NULL
)
}
\arguments{
\item{data}{data.frame; originally generated by \code{\link[=RNAimport]{RNAimport()}}.}

\item{group}{character; vector contains the condition
(ie. treatment or control) of each sample in the order samples are shown in
the \code{data} object from left to right.}

\item{method}{character; method to undertaken differential analysis, choose
from methods of either \link[DESeq2:DESeq]{DESeq2::DESeq} or \link[edgeR:edgeR-package]{edgeR::edgeR}.
Must be stated as either "DESeq2" or "edgeR" in the function.}

\item{dispersionValue}{numeric; manual setting of dispersion value
which is recommended for analysis in experiments without biological
replicates when utilising the \link[edgeR:edgeR-package]{edgeR::edgeR} method.}
}
\value{
Undertakes differential analysis, based on a specified method, and
adds the results to the supplied dataframe. This includes:
\itemize{
\item Count mean
\item Log fold change
\item p-value
\item Adjusted p-value
\item Log counts per million (CPM/RPM)
\item Comparison order
}
}
\description{
This function allows you to compute the differential analysis
of sRNA dicer-derived clusters or mRNA. The function allows the choice
between analysis with \code{DESeq2} or \code{edgeR}.
}
\details{
The analysis allows the users to choose the method which best suits
their data. Notably, \code{DESeq2} cannot compute the analysis when there only
one replicate per condition, but, \code{edgeR} can. Simply set a suitable
dispersion value, based on similar data, to use this feature. The dispersion
value is other wise known as the common Biological squared coefficient
of variation. A typical dispersion value is 0.4 for human data sets, 0.1 for
data on genetically identical model organisms or 0.01 for technical
replicate. See the User’s Guide for the \code{edgeR} package for more details,
\link[edgeR:edgeR-package]{edgeR::edgeR}.
}
\examples{
# load data 
data("sRNA_data_dicercall")

controls <- c("selfgraft_1", "selfgraft_2", "selfgraft_3")

reorder_df <- RNAreorder(sRNA_data_dicercall, controls)

# sample conditions.
groups <- c("Selfgraft", "Selfgraft", "Selfgraft", 
           "Heterograft", "Heterograft", "Heterograft")


## Differential analysis: DEseq2 method
sRNA_DESeq2 <- RNAdifferentialAnalysis(data = sRNA_data_dicercall,
                             group = groups,
                             method = "DESeq2" )


## Differential analysis: edgeR method
sRNA_edgeR <- RNAdifferentialAnalysis(data = sRNA_data_dicercall,
                            group = groups,
                            method = "edgeR" )
}
