% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAdistribution.R
\name{RNAdistribution}
\alias{RNAdistribution}
\title{Plot the distribution of sRNA lengths/classes}
\usage{
RNAdistribution(
  data,
  samples = NULL,
  style,
  facet = TRUE,
  facet.arrange = 3,
  colour = "black",
  together = TRUE,
  relative = FALSE
)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAimport]{RNAimport()}}, can also plot
information generated by \code{\link[=RNAdicercall]{RNAdicercall()}}.}

\item{samples}{character; stores names of samples to analyse and plot.
Argument is required for plotting individual sample replicates, either
individually, overlapped together or in a facet. Use the sample
replicate names present in the data frame, select samples you wish
to plot. Is not required when plotting the sRNA consensus using the argument
\code{total=TRUE}.}

\item{style}{character; \code{style="line"}, \code{style="bar"} or
\code{style="consensus"}. Instructs how to plot the data. Where
\code{style="line"} plots a line graph. Where \code{style="bar"} plots a bar
graph. Where \code{style="consensus"} produces the line graph for the
sRNA dicer-derived consensus generated by the function
\code{\link[=RNAdicercall]{RNAdicercall()}}.}

\item{facet}{logical; forms a matrix of panels defined by row and column
faceting variables. It plots the results for each sample as a bar-chart
and contains it within a single plot. The number of rows in the facet can
be changed using the argument \code{facet.arrange}.
Default \code{facet = TRUE} , plots each sample separately when
\code{facet = FALSE} .}

\item{facet.arrange}{numeric; value supplied to define the number of columns
to include in the facet. This argument is piped into the \code{ ncol}  argument in
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to define the number of columns. By default, this is
set to 3.}

\item{colour}{character; Default colour is "black".}

\item{together}{logical; generates a single line graph, containing all
sample replicate information. Default \code{together=TRUE}.}

\item{relative}{logical; used in conjunction with  \code{style="consensus"}.
Instructs plotting of the relative frequency of all sRNA classes across the
data defined by the consensus dicer-derived sRNA column (see
\code{\link[=RNAdicercall]{RNAdicercall()}} function for more information).}
}
\value{
The function returns a list containing the results in the form of a
data frame and the plot(s). To access an element, simply use the "$" symbol,
and the elements "data" and "plot" will appear. The \code{samples} argument allows
uses to plot specific samples in a single plot (facet bar plot or line graph).
This can encourage closer comparison between sample replicates.
}
\description{
\code{RNAdistribution} plots the distribution of dicer-derived
sRNA classes across samples or across the sRNA consensus
determined by the function \code{\link[=RNAdicercall]{RNAdicercall()}}.
}
\details{
The function can be used to plot a variety of different comparisons and plots.
It can be used to plot the distribution of sRNA classes within each sample
replicate, which can be represented as a bar chart \code{style="bar"} or a
line graph \code{style="line"}. These plots can be represented individually or
in a single plot facet \code{facet="TRUE"} by default. While the option
\code{style="consensus"}, plots the distribution of the consensus
dicer-derived sRNA classes determined by the \code{\link[=RNAdicercall]{RNAdicercall()}} function

To plot the sRNA dicer-derived clusters identified in each sample, the
function extracts the information from the sRNA data set and calculates the
total number of each RNA class identified within a sample, for all samples.

Alternatively, the function allows you to plot the line graph for each sample
together, overlapped on a single graph \code{together=TRUE}. This is not an
option for bar plots.
}
\examples{
# load data 
data('sRNA_data')

p1 <- RNAdistribution(data = sRNA_data, style = "line")

p2 <- RNAdistribution(data = sRNA_data, style = "line", together =FALSE )

p3 <- RNAdistribution(data = sRNA_data, style = "bar")

p3.2 <- RNAdistribution(data = sRNA_data, style = "bar",
                       samples = c("heterograft_1", "heterograft_2",
                       "heterograft_3"))

p4 <- RNAdistribution(data = sRNA_data, style = "bar", facet = FALSE)

p5 <- RNAdistribution(data = sRNA_data, style = "bar",
                      facet = TRUE, facet.arrange = 2 )

# calculate data with dicer-derived consensus :
conditions <- c("heterograft_1", "heterograft_2", "heterograft_3")

# Run function to define sRNA class for each cluster.
sRNA_data_dicercall <- RNAdicercall(data = sRNA_data,
                                  conditions = conditions,
                                  tidy=TRUE)

p6 <- RNAdistribution(data = sRNA_data_dicercall, style = "consensus")

}
