% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAdistribution.R
\name{RNAdistribution}
\alias{RNAdistribution}
\title{Extract the distribution of RNA classes within each sample}
\usage{
RNAdistribution(
  data,
  samples = NULL,
  style = c("bar", "line"),
  facet = TRUE,
  facet.arrange = 3,
  colour = NULL,
  together = TRUE
)
}
\arguments{
\item{data}{a dataframe, on which one of the following functions has already
been called: [RNAimport()],[RNAconsensus()].}

\item{samples}{character vector. Store names of samples to analyse and plot.
Functionality is dependent on the use of argument `facet` for a bar chart
style plot and `together` for a line graph style plot.}

\item{style}{plotting option to choose the style of a line graph or bar chart
to represent your data. Where `style="line"` a line graph will be produced,
while `style=bar` produces a bar graph.}

\item{facet}{Logical; forms a matrix of panels defined by row and column faceting
variables. It plots the results for each sample as a bar-chart and contains
it within a single plot. The number of rows in the facet can be changed using
the argument ` facet.arrange` . Default ` facet = TRUE` , plot each
sample separately when ` facet = FALSE` .}

\item{facet.arrange}{numeric sent to the ` ncol`  argument in
[facet_wrap()] to define the number of columns.}

\item{colour}{bar plot fill colour. Default colour is "darkblue".}

\item{together}{Logical; forms a single line graph with multiple lines each
to represent the sample replicates. Default `together=TRUE`.}
}
\value{
Returns a list containing the data frame and the plot(s). To access
one of the elements simply use the "$" symbol, and the elements "data" and
"plot" will appear. The `samples` argument allows uses to plot specific
samples in a single plot (facet bar plot or line graph). This can encourage
closer comparision between sample replicates.
}
\description{
` RNAdistribution`  extracts information from the RNA summary
data and calculates the total number of each RNA class identified within a
sample, for all samples. The results are plotted as either bar chart, with
an option to facet all plots into a single plot, or as a line graph.
}
\examples{
data('sRNA_data')

p1 <- RNAdistribution(data = sRNA_data, style = "line")

p1.2 <- RNAdistribution(data = sRNA_data, style = "line",
                        samples = c("TomEgg_1", "TomEgg_2", "TomEgg_3"))
p2 <- RNAdistribution(data = sRNA_data, style = "line", together =FALSE )

p3 <- RNAdistribution(data = sRNA_data, style = "bar")

p3.2 <- RNAdistribution(data = sRNA_data, style = "bar",
                       samples = c("TomEgg_1", "TomEgg_2", "TomEgg_3"))

p4 <- RNAdistribution(data = sRNA_data, style = "bar", facet = FALSE)

p4 <- RNAdistribution(data = sRNA_data, style = "bar",
                      facet = FALSE, facet.arrange = 2 )

}
