% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAfeatures.R
\name{RNAfeatures}
\alias{RNAfeatures}
\title{Distribution of sRNA cluster loci across genomic features}
\usage{
RNAfeatures(
  data,
  annotation,
  repeats = NULL,
  promoterRegions = 1000,
  percentage = TRUE
)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAimport]{RNAimport()}}}

\item{annotation}{path; URL, connection or GFFFile object. A genome
reference annotation file (.gff/.gff1/.gff2/.gff3).}

\item{repeats}{path; URL, connection or GFFFile object. A genome
reference annotation file, which specifically contains information on repeat
sequences in the genome (.gff/.gff1/.gff2/.gff3). By default, this is not
required, however if there is a specific repeats annotation file for the
genome it is suggested to supply it.}

\item{promoterRegions}{numeric; define promoter region upstream of genes.
Default is \code{promoterRegions=1000} , ie promoters set at 1-kb upstream of
genes}

\item{percentage}{returns results as a percentage of the total, when
\code{percentage = TRUE} (default). While \code{percentage = FALSE}, results
are returned as a count value representing the number of sRNA clusters
that overlap with a given genomic feature.}
}
\value{
Returns a table containing the number or percentage of overlaps in
the supplied dataset with specific regions in the genome annotation such
as genes, repeats, introns, exons.
}
\description{
Calculates the absolute or relative number of sRNA clusters
which overlap with genomic features, including promoter regions, repeat
regions, exons, introns, and 3'/'5 untranslated regions.
}
\details{
The function calculates the number or percentage of sRNA dicer-derived
clusters which overlap with genomic features based on their genomic
coordinates.

This function can be utilised at different steps in your analysis, but is
particularly useful when identifying whether identified sRNAs meet
expectations and to identify any patterns in the loci which produce mobile
sRNAs.
}
\examples{
data("sRNA_data")
features <- RNAfeatures(data = sRNA_data,
                       annotation = system.file("extdata",
                       "reduced_chr2_Tomato.gff", package="mobileRNA"))
                       
}
\seealso{
\code{\link[=RNAmergeAnnotations]{RNAmergeAnnotations()}} to merge 2 GFF files into 1.
}
