% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAimport.R
\name{RNAimport}
\alias{RNAimport}
\title{Import and organise sRNAseq data sets}
\usage{
RNAimport(input = c("sRNA"), directory, samples, tidy = TRUE, FPKM = FALSE)
}
\arguments{
\item{input}{string; define type of Next-Generation Sequencing dataset.
"sRNA" is the only valid inputs at this time.}

\item{directory}{path; directory containing of sample folders generated by
\code{ShortStack}}

\item{samples}{character; vector naming samples correlating
to outputted folders within the \code{directory} path.}

\item{tidy}{logical; tidy-up dataframe by removing sRNA clusters
where there are zero counts across all replicates.}

\item{FPKM}{logical; calculate the FPKM for each sample}
}
\value{
A dataframe where rows represent sRNA clusters and columns represent
replicate information. Replicate information includes Dicercall, Counts,
and MajorRNA sequence. Each replicate information is distinguishable as
the replicate name is joined as a suffix to each column name. For example, for
a sample called "Sample1", the columns will include DicerCall_Sample1,
Count_Sample1, MajorRNA_Sample1 and RPM_Sample1

The breakdown of each column:
\itemize{
\item \code{Locus} : sRNA cluster locus
\item \code{Chr} : Chromosome
\item \code{start} : start coordinate of cluster
\item \code{end} : end coordinate of cluster
\item \code{Cluster} : name of cluster
\item \code{DicerCall_} : the size in nucleotides of most abundant sRNA in the cluster
\item \code{Counts_} :  number of uniquely aligned sRNA-seq reads that overlap the locus
\item \code{MajorRNA_} : RNA sequence of the most abundant sRNA in the cluster
\item \code{RPM_} : reads per million
\item \code{FPKM_} : Fragments Per Kilobase of transcript per Million mapped reads (only if option activated)
}
}
\description{
Load and organise either sRNAseq results into a
single dataframe containing all experimental replicates specified where rows
represent either a sRNA locus or gene, respectively.
}
\details{
The \code{RNAimport()} function requires the user to supply a directory path and
a character vector. The path must be to the mapping/alignment output.
Following the \code{mobileRNA} method, this path will be to the
\verb{2_alignment_results} folder produced by the \code{\link[=sRNAmapper]{sRNAmapper()}}
function. The \code{RNAimport()} function requires a vector storing the names of
the replicates in the analysis. These should represent and mirror the names
of the folders stored within the \verb{2_alignment_results} folder.

Together this information allows the function to extract the information
stored in "Result.txt" files of each sample.
}
\examples{
\dontrun{

# import sRNAseq data
df_sRNA <- RNAimport(input = "sRNA",
                     directory = "./analysis/sRNA_mapping_results",
                     samples = c("heterograft_1", "heterograft_2",
                     "heterograft_3","selfgraft_1" , "selfgraft_2" ,
                     "selfgraft_3"))


}
# The output of this function can be explore in the data object sRNA_data
data("sRNA_data")
head(sRNA_data)


}
\references{
ShortStack (https://github.com/MikeAxtell/ShortStack)
}
