% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAimport.R
\name{RNAimport}
\alias{RNAimport}
\title{Sample Data Table}
\usage{
RNAimport(results, samples, clusters, totalNumReads, features)
}
\arguments{
\item{results}{Local directory location of sample folders. NOTE: Following
the pre-processing steps, these can be found in second alignment folders.
It is recommended to use results from the unique alignment found in
\code{"./second_alignment_unique/"} over multimapped reads.
It is recommended to run a parallel analysis one using the uniquely
mapped reads and the other using multimapped reads.}

\item{samples}{Vector of characters naming the sample names correlating
to outputted folders either \code{"second_alignment_unique"} or
\code{"second_alignment_multi"}.}

\item{clusters}{Import \code{.gff3} from \code{"reference"} folder.}

\item{totalNumReads}{Table containing containing the total number of reads
for each samples. Each value labeled with the sample name.}

\item{features}{Path to a .bed file containing the genomic information
(annotations/features) which overlaps with the locations of the
dicer-derived clusters. The file is created as a part of the pre-processing
steps, see fun[find_overlap] function for more information.}
}
\value{
A large dataframe with data for each sample including Dicercall,
Count and Cluster ID.
}
\description{
Import specific data from the \code{"Results"} file of each
sample and collate into a single dataframe. Plus, calculates the \code{FPKM}
 for each sample.
}
\examples{
\dontrun{
genomic_features <- "./data/reference/overlap.bed"
directory <- "./data/alignment_unique_two/"
sample_names <- c("TomEgg_1","TomEgg_2", "TomEgg_3",
              "TomTom_1", "TomTom_2", "TomTom_3" )

clusters <- rtracklayer::import.gff("./data/reference/ClustersInfo.gff3")

total_reads <- c(24441759,21378845, 21482356, 3951725, 3343954, 2586910)
names(total_reads) <- samples
siRNA_data <- RNAimport(results = directory,
                           samples = sample_names,
                           clusters = clusters,
                           totalNumReads = total_reads
                           features = genomic_features)
}

}
