% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAimport.R
\name{RNAimport}
\alias{RNAimport}
\title{Import and organise sRNAseq data set}
\usage{
RNAimport(loci, directory, samples, report = TRUE, features = NULL)
}
\arguments{
\item{loci}{dataframe; rows represent individual sRNA dicer-derived clusters
columns state genomic location (labelled "Locus") and the cluster name
(labelled "Cluster"). This file contains an accululation of all
sRNA dicer-derived clusters  identified across samples in the analyse, without
duplication. This can be produced using the [RNAlocate::RNAloci()]
function}

\item{directory}{Path to directory containing of sample folders. NOTE: Following
the suggested pre-processing steps, these can be found in second alignment
folders.}

\item{samples}{Vector of characters naming the sample names correlating
to outputted folders located in the \code{`directory`} argument path.}

\item{report}{Logical; prompts progress.}

\item{features}{Path to a txt file containing the genomic information
(annotations/features) which overlaps with the locations of the
dicer-derived clusters. The file is created as a part of the pre-processing
steps, see [RNAlocate::findOverlap()] function for more information. This is
an optional argument, but will streamline your analysis later on if you wish
to identify the genomic location of the cluster and the type of feature it
overlaps with.}
}
\description{
Organises the sRNAseq results data, outputted by ShorStack, for
each sample replicate into a single dataframe using the loci information
produced by the [RNAlocate::RNAloci()] function which contains a
list of all the clusters identified across samples.
}
\details{
Supply the directory to the folder storing the various sample replicate folders
produced by ShortStack analysis. Within each sample replicate folder, there is
"Result.txt". This file contains many different columns, but
for the analysis the columns of interest are  "Locus", "DicerCall", "Reads", and
"RPM" which will all be imported and loads into the new data frame. Note that
"Reads" will be converted to "Counts" as output of the function.
Locus contains the genomic locations of clusters,
Dicercall contains the most likely size of the cluster, Reads contains the
number of reads mapped to the cluster (recommended to use unique mapping,
hence these will be the number of uniquely mapped reads, ie, not including
multimapped reads). Lastly, RPM contains the Reads per Million score.

\code{"RNAimport"} takes the results from  pre-processed sRNAseq data which
has been mapped and undertaken cluster analysis through ShortStack, organises
 into a single data frame. Each row represent an sRNA dicer-derived cluster
 within the analysis and columns 1 - 5 supply information on the cluster
 including the locus, the separated coordinates
(chromosome, start, end) and cluster name.

Further columns represent data imported for each samples including DicerCall,
Counts and RPM. The DicerCall represents the size of most
abundant small RNA size based on the parameter used in ShortStack.
The Count column represents the number of aligned sRNA-seq reads that overlap
the locus. The RPM represents the reads per million.
For each replicate included in the analysis, these columns are labeled with
the type and then then name of the sample, for example, for
a sample called "Sample1", the information from this sample will be stored in
columns DicerCall_Sample1, Count_Sample1 and RPM_Sample1.
}
\examples{
\dontrun{
genomic_features <- "./data/reference/overlap.bed"

clusters <- utils::read.table(file = "./data/reference/ClustersInfo.txt",
header = TRUE, sep = "\t", stringsAsFactors = TRUE,comment.char="")

RNAorganise(loci = clusters,
           directory = "./replicates_EggTom_spike/",
           samples = c( "TomEgg_1", "TomEgg_2", "TomEgg_3",
                       "TomTom_1" , "TomTom_2" , "TomTom_3"),
           features = genomic_features)


}

}
