% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmean.R
\name{RNAmean}
\alias{RNAmean}
\title{Calculate mean FPKM across samples}
\usage{
RNAmean(data, conditions)
}
\arguments{
\item{data}{data frame, see [RNAlocate::RNAimport()] to produce an organised
object of sample data.}

\item{conditions}{Character vector; represent sample names.}
}
\value{
A data frame containing all existing columns in the input data object,
plus, an additional columns. One containing the FPKM means, and Count means.

As default, the FPKM means will be stored in the column named `mean_FPKM`,
while the count means are stored in column names `mean_Count`.
}
\description{
Using the data, the function calculates the mean FPKM for the
selected samples and adds the results to the working data frame.
}
\examples{
data("sRNA_data")
selected_samples <- c("TomEgg_1", "TomEgg_2", "TomEgg_3")
means <- RNAmean(data = sRNA_data, conditions = selected_samples)

}
