% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmergeAnnotations.R
\name{RNAmergeAnnotations}
\alias{RNAmergeAnnotations}
\title{Merge two GFF genome annotation files}
\usage{
RNAmergeAnnotations(
  annotationA,
  annotationB,
  out_dir,
  abbreviationAnnoA = "A",
  abbreviationAnnoB = "B",
  replace_chr_names = TRUE
)
}
\arguments{
\item{annotationA}{the path  \code{base::connection()} to a genome annotation
file in GFF3 format}

\item{annotationB}{the path  \code{base::connection()} to a genome annotation
file in GFF3 format}

\item{out_dir}{either a character string or a \code{base::connections()} open
for writing. Place path to output directory in "", including file output name
with extension of either \code{.gff} or \code{.gff3}.}

\item{abbreviationAnnoA}{a string placed in "", to replace chromosome names
within \code{annotationA}.Default set as "A".}

\item{abbreviationAnnoB}{a string placed in "", to replace chromosome names
within \code{annotationB}.Default set as "B".}

\item{replace_chr_names}{a logical value indicating whether the chromosome
names of the supplied annotation files are to be altered or not. As default,
chromosome names are altered.}
}
\value{
The function output the individually altered annotation files, plus, a merged
annotation file made up of the two altered files. The altered annotation files
contain either default or custom alterations to the chromosome names and are
written to the original file connection with the addition of "_altered.gff3"
at the end of the file name.

The  merged annotation file is outputted to the global object variable when
assigned, as well as writing the merged annotation file in
GFF format to a give output directory.
}
\description{
Merges two genomes annotation files (\verb{GFF format}) into one
single GFF file saved to desired directory.
}
\details{
The functions primary goal is to merge two GFF files, however, when
merging genomic files it is critical that the two genomes are distinguishable
by the chromosome names. As a default setting, the function extracts the
chromosome names for a given GFF file and adds an ID before each chromosome
name and removes any periods.

The function requires the input of two reference annotation genomes, where
one represents \code{annotationA} and the other represents \code{annotationB}.
As default, the function will rename the chromosome names in \code{annotationA}
to "A". For example, A0, A1, A2 etc. To set a custom chromosome name for
\code{annotationA} alter the argument \code{abbreviationAnnoA}. While, for
\code{annotationB} as default the chromosome names will be named "B", for example,
B0, B1, B2 etc. To set a custom chromosome name for \code{annotationB} alter the
argument \code{abbreviationAnnoB}. The function can do so by  draw the
chromosome number within the given GFF file, remove all prior character or
numerical values, and replace it with the supplied string.

IMPORTANT:  The genome reference and annotation of a species
must have chromosomes with matching names. It is critical that if you used
the \code{\link[=RNAmergeGenomes]{RNAmergeGenomes()}} function to to create a merged reference
genome,that you treat the input annotations in the same way.
}
