% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmergeAnnotations.R
\name{RNAmergeAnnotations}
\alias{RNAmergeAnnotations}
\title{Merge two GFF genome annotation files}
\usage{
RNAmergeAnnotations(
  annotationA,
  annotationB,
  out_dir,
  abbreviationAnnoA = "A",
  abbreviationAnnoB = "B"
)
}
\arguments{
\item{annotationA}{GRanges object; a genome annotation assembly file in
GFF format.}

\item{annotationB}{GRanges object; a genome annotation assembly file in
GFF format.}

\item{out_dir}{either a character string or a \code{base::connections()} open
for writing. Place path to output directory in "", including file output name
with extension of either \code{.gff} or \code{.gff3}.}

\item{abbreviationAnnoA}{a string placed in "", to replace chromosome names
within \code{annotationA}.Default set as "A".}

\item{abbreviationAnnoB}{a string placed in "", to replace chromosome names
within \code{annotationB}.Default set as "B".}
}
\value{
The function output the individually altered annotation files, plus, a merged
annotation file to the global environment and the given directory.
The altered annotation files contain either default or custom alterations to
the chromosome names and are written to the same directory as the merged file,
with the given names addition of \code{annotationA_altered.gff} and
\code{annotationB_altered.gff}, respectively.
}
\description{
Merges two genomes annotation files (\verb{GFF format}) into one
single GFF file saved to desired directory. Adds prefix to each chromosome
name in the genome annotations to ensure each assembly is distinguishable
within the merged file, and saved the individually altered references to
local directory.
}
\details{
The functions primary goal is to merge two GFF files, however, when
merging genomic files it is critical that the two genomes are distinguishable
by the chromosome names. As a default setting, the function extracts the
chromosome names for a given GFF file and adds an ID before each chromosome
name and removes any periods.

As default, the function will rename the chromosome names in \code{annotationA}
to "A". For example, A0, A1, A2 etc. To set a custom chromosome name for
\code{annotationA} alter the argument \code{abbreviationAnnoA}. While, for
\code{annotationB} as default the chromosome names will be named "B", for example,
B0, B1, B2 etc. To set a custom chromosome name for \code{annotationB} alter the
argument \code{abbreviationAnnoB}.

The merged genome is saved to the specified output directory, and requires
the user to set the name with a \code{gff} extension. The user must load the
annotations into the R Global Environment, hence, the alterations made by
this function will alter the variable. The altered annotations will also
be saved to the same directory as supplied for the output file, with a fixed
output name of \code{annotationA_altered.gff} and \code{annotationB_altered.gff},
respectively.

IMPORTANT:  The genome reference and annotation of a species
must have chromosomes with matching names. It is critical that if you used
the \code{\link[=RNAmergeGenomes]{RNAmergeGenomes()}} function to to create a merged reference
genome,that you treat the input annotations in the same way.
}
\examples{


 # Read in GFF files
anno1 <- rtracklayer::import(system.file("extdata",
"chr12_Eggplant_V4.1_function_IPR_final.gff.gz", package="mobileRNA"),
format = "gff")

anno2 <- rtracklayer::import(system.file("extdata",
"chr2_ITAG4.0_gene_models.gff.gz", package="mobileRNA"),
format = "gff")

# edit and merge
merged_anno <- RNAmergeAnnotations(annotationA = anno1,
                                 annotationB = anno2,
             out_dir = "./merged_annotation.gff3")

## Set specific pre-fixes:
###  annotationA represents Solanum melongena, so abbreviated to `SM`
###  annotationB represents Solanum lycopersicum, so abbreviated to `SL`

merged_anno_2 <- RNAmergeAnnotations(annotationA = anno1,annotationB = anno2,
              out_dir = "./merged_annotation.gff3",
              abbreviationAnnoA = "SM",abbreviationAnnoB = "SL")





}
