% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmergeGenomes.R
\name{RNAmergeGenomes}
\alias{RNAmergeGenomes}
\title{Merge two FASTA genome assemblies}
\usage{
RNAmergeGenomes(
  genomeA,
  genomeB,
  out_dir,
  abbreviationGenomeA = "A",
  abbreviationGenomeB = "B"
)
}
\arguments{
\item{genomeA}{a large DNAStringSet; a genome reference assembly file in
FASTA format (.fa/.fasta).}

\item{genomeB}{a large DNAStringSet; a genome reference assembly file in
FASTA format (.fa/.fasta).}

\item{out_dir}{either a character string or a \code{base::connections()} open
for writing. Place path to output directory in "", including file output name.
Output name must have a ".fa" extension.}

\item{abbreviationGenomeA}{a string placed in "", to replace chromosome names
within \code{genomeA}.Default set as "A".}

\item{abbreviationGenomeB}{a string placed in "", to replace chromosome names
within \code{genomeB}.Default set as "B".}
}
\value{
Returns a single FASTA format file containing both  genome assemblies
with edited chromosome names (prefixes, and removal of periods) to the give
directory, as well as the individual edited genomes assemblies named
"genomeA_altered.fa" and "genomeB_altered.fa", respectively.
}
\description{
Merges two reference genomes (.fa/.fasta). into one single
reference with modified chromosome names to ensure distinguishability.
}
\details{
The functions primary goal is to merge two FASTA files, however, when
merging genomic files it is critical that the two genomes are distinguishable
by the chromosome names. As a default setting, the function extracts the
chromosome names for a given FASTA file and alters the name while retaining
the identifying number.

The function requires the input of two reference genomes, where one
represents \code{Genome-A} and the other represents \code{Genome-B}. As default, the
function will rename the chromosome names in \code{Genome-A} to "A". For example,
A0, A1, A2 etc. To set a custom chromosome name for \code{Genome-A} alter the
argument \code{abbreviationGenomeA}. While, for  \code{Genome-B} as default the
chromosome names will be named "B", for example, B0, B1, B2 etc. To set a
custom chromosome name for \code{Genome-A} alter the argument
\code{abbreviationGenomeB}. The function can do so by  draw the chromosome
number within the given GFF file, remove all prior character or numerical
values, and replace it with the supplied string.

IMPORTANT:  The genome reference and annotation of a species
must have chromosomes with matching names. It is critical that if you used
the \code{\link[=RNAmergeAnnotations]{RNAmergeAnnotations()}} function to to create a merged genome
annotation,that you treat the input references in the same way.
}
\examples{

# URL for example genome assemblies
 url_remote <- "https://github.com/KJeynesCupper/assemblies/raw/main/"

 fasta_1 <- paste0(url_remote, "chr12_Eggplant_V4.1.fa.gz")

 fasta_2 <- paste0(url_remote,"chr2_S_lycopersicum_chromosomes.4.00.fa.gz")


 # load into R
 ref1 <- Biostrings::readDNAStringSet(fasta_1,format = "fasta")
 ref2 <- Biostrings::readDNAStringSet(fasta_2,format = "fasta")

# run function to merge
merged_ref <- RNAmergeGenomes(genomeA = ref1, genomeB = ref2,
                               out_dir = "./merged_ref.fa")




## or, to set specific changes to chromosome names:
 # genomeA represents Solanum melongena and the chromosomes will be
 # abbreviated to `SM.
# genomeB represents Solanum lycopersicum and the chromosomes will be
# abbreviated to `SL`.

merged_ref2 <- RNAmergeGenomes(genomeA = ref1,
            genomeB = ref2 ,
            out_dir = "./merged_ref.fa",
            abbreviationGenomeA = "SM",
            abbreviationGenomeB = "SL")

}
