% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmergeGenomes.R
\name{RNAmergeGenomes}
\alias{RNAmergeGenomes}
\title{Merge two FASTA genome assemblies}
\usage{
RNAmergeGenomes(
  genomeA,
  genomeB,
  out_dir,
  abbreviationGenomeA = "A",
  abbreviationGenomeB = "B",
  cores = TRUE,
  number_cores = 2
)
}
\arguments{
\item{genomeA}{path; directory path to a genome reference assembly file in
FASTA format (.fa/.fasta).}

\item{genomeB}{path; directory path to a genome reference assembly file in
FASTA format (.fa/.fasta).}

\item{out_dir}{either a character string or a \code{base::connections()} open
for writing. Place path to output directory in "", including file output name.
Output name must have a ".fa" extension.}

\item{abbreviationGenomeA}{a string placed in "", to replace chromosome names
within \code{genomeA}.Default set as "A".}

\item{abbreviationGenomeB}{a string placed in "", to replace chromosome names
within \code{genomeB}.Default set as "B".}

\item{cores}{logical;  the number of CPU cores for parallel computation.
By default, \code{cores = TRUE} which tells the system to detect the number of CPU
cores on the current host.While, \code{cores = FALSE} allows a user defined number
of cores which is stored in the variable \code{number_cores}.}

\item{number_cores}{numeric; number of CPU cores to use during parallel
computation.}
}
\value{
Returns a single FASTA format file containing both  genome assemblies
with edited chromosome names (prefixes, and removal of periods) to the give
directory.
}
\description{
Merges two reference genomes (.fa/.fasta). into one single
reference with modified chromosome names to ensure distinguishability.
}
\details{
Please be aware that this function will take a very long time to process if
you are working with large genomes which together take up more than 1Gb
storage. Please be patient and allow the function to run.

The functions primary goal is to merge two FASTA files, however, when
merging genomic files it is critical that the two genomes are distinguishable
by the chromosome names. As a default setting, the function extracts the
chromosome names for a given FASTA file and alters the name while retaining
the identifying number.

The function requires the input of two reference genomes, where one
represents \code{Genome-A} and the other represents \code{Genome-B}. As default, the
function will rename the chromosome names in \code{Genome-A} to "A". For example,
A0, A1, A2 etc. To set a custom chromosome name for \code{Genome-A} alter the
argument \code{abbreviationGenomeA}. While, for  \code{Genome-B} as default the
chromosome names will be named "B", for example, B0, B1, B2 etc. To set a
custom chromosome name for \code{Genome-A} alter the argument
\code{abbreviationGenomeB}. The function can do so by  draw the chromosome
number within the given GFF file, remove all prior character or numerical
values, and replace it with the supplied string.

Please note that this function uses parallel computation using the packages
(\code{parallel}, \code{foreach} & \code{doParallel}) to improve the processing speed. Hence
common errors may occur is you have some parallel computing going on in the
background. To improve the handling, the function will create a number of
temporary files in your output directory, which are then merged into a single
file and deleted.

IMPORTANT:  The genome reference and annotation of a species
must have chromosomes with matching names. It is critical that if you used
the \code{\link[=RNAmergeAnnotations]{RNAmergeAnnotations()}} function to to create a merged genome
annotation,that you treat the input references in the same way.
}
\examples{

# URL for example genome assemblies
 url_remote <- "https://github.com/KJeynesCupper/assemblies/raw/main/"

 fasta_1 <- paste0(url_remote, "chr12_Eggplant_V4.1.fa.gz")

 fasta_2 <- paste0(url_remote,"chr2_S_lycopersicum_chromosomes.4.00.fa.gz")


# run function to merge
merged_ref <- RNAmergeGenomes(genomeA = fasta_1, genomeB = fasta_2,
                               out_dir = tempfile("merged_genome",tempdir(), fileext = ".fa"), 
                               cores = FALSE,number_cores = 1)




## or, to set specific changes to chromosome names:
 # genomeA represents **Solanum melongena** and the chromosomes will be
 # abbreviated to `SM`.
# genomeB represents **Solanum lycopersicum** and the chromosomes will be
# abbreviated to `SL`.

merged_ref2 <- RNAmergeGenomes(genomeA = fasta_1,genomeB = fasta_2 ,
            out_dir = tempfile("merged_genome",tempdir(), fileext = ".fa"),
            abbreviationGenomeA = "SM",
            abbreviationGenomeB = "SL",  
           cores = FALSE, number_cores = 1)
            
}
