% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmobile.R
\name{RNAmobile}
\alias{RNAmobile}
\title{Identify potential mobile sRNA molecules}
\usage{
RNAmobile(
  data,
  controls,
  genome.ID,
  task = NULL,
  statistical = FALSE,
  alpha = 0.1,
  threshold = NULL
)
}
\arguments{
\item{data}{Numeric data frame}

\item{controls}{Character vector; containing names of control samples.}

\item{genome.ID}{a character string related to the chromosomes in a
particular genome. A distinguishing feature of the genome of interest or
non-interest in the chromosome name (\code{chr} column).}

\item{task}{an option to keep or remove the chromosomes containing the
identifying string. To keep the chromosomes with the ID, set task=keep.
To remove, set \code{task="remove"}. As default, task is set to \code{keep}.}

\item{statistical}{If TRUE, will undertake statistical filtering based on the
a p-value or adjusted p-value threshold stated by \code{alpha}.Default
set at FALSE. Requires presence of columns containing statistical data.
In order to filter by the adjusted p-value, a column named \code{padjusted} must
be present. See \code{\link[=RNAanalysis]{RNAanalysis()}} to calculate statistical values.}

\item{alpha}{numeric; adjusted p-value cutoff as the target FDR for
independent filtering. Default is 0.1. Only mobile molecules with adjusted
p-values equal or lower than specified are returned.}

\item{threshold}{numeric; set a threshold level. For sRNAseq, this represents
filtering by the minimum number of replicates that defined the
consensus dicercall which is stored in the \code{DicerCounts} column.}

\item{input}{character; must be either "sRNA" or "mRNA" to represent the type
of data, required when setting threshold.}
}
\value{
A data-frame containing candidate mobile sRNAs, which could be
further filtered based on statistical significance and the ability to
by-pass the threshold which determines the number of replicates that
defined the consensus dicercall.
}
\description{
A function to identify potential mobile sRNA molecules
detected in a tissue which originated from a different genotype, usually
related to a chimeric system.
}
\details{
The function identifies candidate mobile sRNAs, by selecting sRNA clusters
mapped to the foreign genome. It does so by either keeping or removing sRNA
mapped to a given genome. To do so, it requires a common pre-fix across
chromosomes of the same genome. See \code{\link[=RNAmergeGenomes]{RNAmergeGenomes()}} for more
information.

A greater confidence in the mobile sRNA candidates can be achieved by setting
a threshold that considers the number of replicates which contributed to
defining the consensus dicercall (ie. consensus sRNA classification). This
parameter filters based on the \code{DicerCounts} column introduced by the
\code{\link[=RNAdicercall]{RNAdicercall()}} function.

\strong{Statistical Analysis}
The function also allows for filtering using statistical inference generated
from the differential analysis of the total dataset using the function
\code{\link[=RNAanalysis]{RNAanalysis()}}. When \code{statistical=TRUE}, the feature is enabled
and selected mobile molecules that meet the adjusted p-value cutoff defined
by \code{alpha}.
}
\examples{


data("sRNA_data_consensus")


# vector of control names
controls <- c("selfgraft_1", "selfgraft_2" , "selfgraft_3")

# Locate potentially mobile sRNA clusters associated to tomato, no
# statistical analysis
mobile_df1 <- RNAmobile(data =  sRNA_data_consensus,
                    controls = controls,
                    genome.ID = "B_",
                    task = "keep",
                    statistical = FALSE)


}
