% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmobile.R
\name{RNAmobile}
\alias{RNAmobile}
\title{Identify potential mobile sRNA molecules}
\usage{
RNAmobile(
  data,
  controls,
  id,
  task = NULL,
  statistical = FALSE,
  padj = 0.05,
  p.value = NULL
)
}
\arguments{
\item{data}{Numeric data frame}

\item{controls}{Character vector; containing names of control samples.}

\item{id}{a character string related to the chromosomes in a particular
genome. A distinguishing feature of the genome of interest or non-interest in
the chromosome name (\code{chr} column).}

\item{task}{an option to keep or remove the chromosomes containing the
identifying string. To keep the chromosomes with the ID, set task=keep.
To remove, set \code{task="remove"}. As default, task is set to \code{keep}.}

\item{statistical}{If TRUE, will undertake statistical filtering based on the
a p-value or adjusted p-value threshold stated by \code{padj} & \code{p.value}.Default
set at FALSE. Requires presence of columns containing statistical data.
In order to filter by the adjusted p-value, a column named \code{padjusted} must
be present. Similarly, to filter by the p-value, a column named \code{pvalue} must
be present. See \code{\link[=RNAanalysis]{RNAanalysis()}} to calculate statistical values.}

\item{padj}{A user defined numeric value to represent the adjusted p-value
threshold to define statistic significance. Defaults set at 0.05.Only mobile
molecules with adjusted p-values equal or lower than specified are returned.}

\item{p.value}{A user defined numeric value to represent the p-value
threshold to define statistic significance. There is no default value, set
this instead of using an adjusted p-value to filter molecules. Only mobile
molecules with p-values equal or lower than specified are returned.}
}
\value{
A data-frame containing statistically significant mobile small
RNA molecules.
}
\description{
A function to identify potential mobile small RNA molecules
traveling from one genotype to another in a hetero-grafted system.
}
\details{
The function undertakes two different roles. First, it selects the sRNA
cluster which are mapped to a particular genome. It does so by either keeping
or removing sRNA mapped to chromosomes. Hence, this function will only work
if the two genomes are distinguishable by their chromosome names.

The second step, after selecting clusters which are mapped to the genome of
interest, removes sRNA clusters which were incorrectly mapped. These
are clusters which have counts or RPM values in the control samples
(ie. same genome as the destination tissue in the hetero-graft condition).
These samples should not have counts if the sRNA originates from a different
genotype to the control.

The function also allows for statistical analysis based on the results
collect from differential analysis of the total dataset using the function
\code{\link[=RNAanalysis]{RNAanalysis()}}. This features enables the filtering of sRNA
clusters which meet a specific p-value or adjusted p-values.
}
\examples{


data("sRNA_data_consensus")


# vector of control names
controls <- c("selfgraft_1", "selfgraft_2" , "selfgraft_3")

# Locate potentially mobile sRNA clusters associated to tomato, no
# statistical analysis
mobile_df1 <- RNAmobile(data = sRNA_data_consensus,
                    controls = controls,
                    id = "SL40",
                    task = "keep",
                    statistical = FALSE)



 # Locate potentially mobile sRNA clusters associated to tomato, include
 # statistical analysis

## undertake statistical analysis with either edgeR or DESeq2, here we use
# # DESeq2
groups <- c("Heterograft", "Heterograft", "Heterograft",
          "Selfgraft", "Selfgraft", "Selfgraft")

analysis_df <- RNAanalysis(data = sRNA_data_consensus,
                             group = groups,
                             method = "DESeq2" )

## locate mobile sRNA using p-adjusted value
mobile_df2 <- RNAmobile(data = analysis_df,
                    controls = controls,
                    id = "SL40",
                    task = "keep",
                    statistical = TRUE)

## or, locate mobile sRNA using p-value value
mobile_df3 <- RNAmobile(data = analysis_df,
                    controls = controls,
                    id = "SL40",
                    task = "keep",
                    statistical = TRUE,
                    p.value = 0.05)



# Locate local sRNA clusters associated to eggplant, include statistical
# analysis
mobile_df4 <- RNAmobile(data = sRNA_data_consensus,
                    controls = controls,
                    id = "SL40",
                    task = "remove",
                    statistical = FALSE)


}
