% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsequences.R
\name{RNAsequences}
\alias{RNAsequences}
\title{Extract sRNA cluster sequences}
\usage{
RNAsequences(data, original = FALSE)
}
\arguments{
\item{data}{dataframe; containing columns starting with \code{Cluster} & \code{MajorRNA}}

\item{original}{logical; output results to working date as additional
columns (\code{original=TRUE}), or output as new dataframe (\code{original=FALSE}).
by default, FALSE}
}
\value{
The results can be stored in the the working dataframe supplied to the
function or stored as a new dataframe containing only the results from the
function. The results includes:
\itemize{
\item Match:  logical; whether the sRNA sequence is consistent across replicates
\item Sequence:  character; sequence of the most abundant sRNA within a cluster
\item Complementary_RNA: character; complementary RNA nucleotide sequence
\item Complementary_DNA: character; complementary DNA nucleotide sequence
\item Width: numeric; length of nucleotide sequence
}
}
\description{
Extract sRNA cluster sequences; identifies whether the most
abundant sRNA for a sRNA cluster id consistent across the replicates, and if
so extracts the sRNA nucleotide sequence.

The function also calculates the RNA & DNA complementary sequences, as well
as stating the length of the sequence.

The results can be stored in the the working dataframe supplied to the
function or stored as a new dataframe containing only the results from the
function. This includes:
\itemize{
\item Match:  logical; whether the sRNA sequence is consistent across replicates
\item Sequence:  character; sequence of the most abundant sRNA within a cluster
\item Complementary_RNA: character; complementary RNA nucleotide sequence
\item Complementary_DNA: character; complementary DNA nucleotide sequence
\item Width: numeric; length of nucleotide sequence
}
}
\examples{

data('sRNA_data_mobile')

mobile_sequences <- RNAsequences(sRNA_data_mobile)

}
