% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsubset.R
\name{RNAsubset}
\alias{RNAsubset}
\title{Subset data to select a specific sRNA population based on dicercall size}
\usage{
RNAsubset(data, class, statistical = FALSE, alpha = 0.05)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAimport]{RNAimport()}}, containing
the additional information generated by \code{\link[=RNAdicercall]{RNAdicercall()}} which
defines the consensus sRNA class for each sRNA dicer-derived clusters}

\item{class}{numeric; sRNA dicercall class(es) to select. Based on size in
nucleotides.}

\item{statistical}{logical; filter and select statistical significant sRNA.
If \code{statistical=TRUE}, sRNA clusters are selected based on
p-adjusted alpha threshold.}

\item{alpha}{numeric; user defined numeric value to represent the adjusted
p-value threshold to define statistic significance. Defaults setting
\code{alpha=0.05}.}
}
\value{
A dataframe containing only sRNA clusters defined with a specific
sRNA dicer-derived consensus size.
}
\description{
Subset the existing dataframe to contain only the desired sRNA
class(s) based on the consensus dicercall determination and statistical
significance.
}
\details{
See \code{\link[=RNAdicercall]{RNAdicercall()}} for information on defining the consensus
sRNA dicercall class for each cluster. The function allows the choice to
filtered the data by statistical significance based on differential expression
analysis, see \code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}}. Set
\code{statistical=TRUE} to filtered by statistical significance (p-adjusted).

It is important to consider what point in your analysis you subset the data
or/and undertake differential analysis to achieve statistical values.
Subsetting the dataset into groups based on the sRNA class will
create a smaller set of data for each to draw statistical differences.
Depending on the size of your data, and analysis aims this should be taken
into consideration.
}
\examples{
data("sRNA_data_dicercall")

# Subset data for  24-nt sRNAs
sRNA_24 <- RNAsubset(sRNA_data_dicercall, class = 24)


# Subset data for 21/22-nt sRNAs
sRNA_2122 <- RNAsubset(sRNA_data_dicercall, class = c(21, 22))

}
