% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsummary.R
\name{RNAsummary}
\alias{RNAsummary}
\title{Summarise the abundance of sRNA clusters}
\usage{
RNAsummary(
  data,
  alpha = 0.1,
  chimeric = FALSE,
  controls = NULL,
  genome.ID = NULL
)
}
\arguments{
\item{data}{data.frame; generated by \code{\link[=RNAdifferentialAnalysis]{RNAdifferentialAnalysis()}}}

\item{alpha}{numeric; user defined numeric value to represent the adjusted
p-value threshold to define statistic significance. Defaults setting
\code{alpha=0.1}.}

\item{chimeric}{logical; state whether system is chimeric: contains multiple
genomes/genotypes.}

\item{controls}{character; vector of control condition sample names.}

\item{genome.ID}{character; chromosome identifier of foreign genome in
chimeric system}
}
\value{
Prints a summary of the sRNA clusters which align with the
adjusted p-value cutoff and states the number which have a positive and
negative log-fold change. Where a positive log-fold change represents an
increase in abundance and a negative value represents a decrease in abundance
between the conditions.
}
\description{
Print a summary of the statistical analysis of sRNA clusters
from a mobileRNA analysis
}
\details{
When working with a chimeric system, for example interspecific grafting,
mapping errors can easily be recognised and eliminated. Here, these can be
eliminated by supplying some extra parameter information. State
\code{chimeric=TRUE} and supply the chromosome identifier of the foreign genome
(ie. not the tissue sample genotype, but the genotype from which any
potential mobile molecules could be traveling from) to the \code{genome.ID}
parameter & the control condition samples names to the \code{controls} parameter.
}
\examples{

# load data 
data("sRNA_data_consensus")

controls <- c("selfgraft_1", "selfgraft_2", "selfgraft_3")

reorder_df <- RNAreorder(sRNA_data_consensus, controls)

# sample conditions.
groups <- c("Selfgraft", "Selfgraft", "Selfgraft", 
           "Heterograft", "Heterograft", "Heterograft")


## Differential analysis: DEseq2 method
sRNA_DESeq2 <- RNAdifferentialAnalysis(data = sRNA_data_consensus,
                             group = groups,
                             method = "DESeq2" )
                          
                   
RNAsummary(sRNA_DESeq2)
                                 

}
