% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sRNA_data}
\alias{sRNA_data}
\title{The \code{sRNA_data} data set}
\usage{
data(sRNA_data)
}
\description{
This data sets represents the results of the pre-processing. The data has
been organised using the \code{sample_table} function.
}
\details{
Tomato (Solanium lycopersicum) and eggplant (Solanium melongena) were used
as grafting partners in a hetero-grafting experiment. The scion was
represented by tomato and the rootstock was represented by the eggplant. The
experiment used a tomato self-graft as a control.

The dataset holds three replicates for the hetero-graft treatment (TomEgg_1,
TomEgg_2, TomEgg_3) and three replicates for the self-graft control
(TomTom_1, TomTom_2, TomTom_3).
}
\keyword{datasets}
