% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sRNA_data}
\alias{sRNA_data}
\title{The \code{sRNA_data} data set}
\usage{
data(sRNA_data)
}
\description{
Simulated sRNAseq dataset
}
\details{
This data sets represents the results of the pre-processing. The data has
been organised using the \code{sample_table} function.

Tomato (Solanium lycopersicum) and eggplant (Solanium melongena) were used
as grafting partners in a hetero-grafting experiment. The scion was
represented by tomato and the rootstock was represented by the eggplant. The
experiment used a tomato self-graft as a control.

The dataset holds three replicates for the hetero-graft treatment (TomEgg_1,
TomEgg_2, TomEgg_3) and three replicates for the self-graft control
(TomTom_1, TomTom_2, TomTom_3).

sRNA_data

Simulates data is taken from eggplant and tomato sRNAseq samples and
created to simulate to movement of sRNA molecules from an Tomato rootstock to
an Eggplant Scion. Three Eggplant replicates were spiked with the same 150
tomato sRNA clusters, and named "heterograft_" 1 to 3. The analysis compares
these heterografts to three Eggplant self-grafts which are the original
un-spiked Eggplant replicates.
@return Dataframe in global environment
@examples
data("sRNA_data")
}
\keyword{dataset}
