% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sRNA_data_consensus}
\alias{sRNA_data_consensus}
\title{sRNA_data_consensus}
\usage{
data(sRNA_data_consensus)
}
\description{
Simulated sRNAseq dataset with RNA consensus
}
\details{
Simulates data is taken from eggplant and tomato sRNAseq samples and
created to simulate to movement of sRNA molecules from an Tomato rootstock to
an Eggplant Scion. Three Eggplant replicates were spiked with the same 150
tomato sRNA clusters, and named "heterograft_" 1 to 3. The analysis compares
these heterografts to three Eggplant self-grafts which are the original un-spiked
Eggplant replicates.
}
\keyword{dataset}
