% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRNAmapper.R
\name{sRNAmapper}
\alias{sRNAmapper}
\title{Map & cluster sRNAseq reads using \code{ShortStack}}
\usage{
sRNAmapper(
  input_files_dir,
  output_dir,
  genomefile,
  condaenv,
  threads = 6,
  mmap = "u",
  dicermin = 20,
  dicermax = 24,
  mincov = 2,
  pad = 75,
  tidy = TRUE
)
}
\arguments{
\item{input_files_dir}{path; directory containing only the FASTQ sRNAseq
samples for analysis. Note that all samples in this directory will be used by
this function.}

\item{output_dir}{path; directory to store output.}

\item{genomefile}{path; path to a FASTA file.}

\item{condaenv}{character; name or directory of the Conda environment to use
where \code{ShortStack} (>4.0) is installed}

\item{threads}{numeric; set the number of threads to use where more threads
means a faster completion time. Default is 6.}

\item{mmap}{character; define how to handle multi-mapped reads. Choose from
"u", "f" or "r". Where "u" means only uniquely-aligned reads are used as
weights for placement of multi-mapped reads. Where "f" means fractional
weighting scheme for placement of multi-mapped reads and "r" mean
multi-mapped read placement is random. Default is "u".}

\item{dicermin}{integer; the minimum size in nucleotides of a valid small
RNA. This option sets the bounds to discriminate dicer-derived small RNA loci
from other loci. Default is 20.}

\item{dicermax}{integer; the maximum size in nucleotides of a valid small
RNA. This option sets the bounds to discriminate dicer-derived small RNA loci
from other loci.  Default is 24.}

\item{mincov}{numeric; minimum alignment depth, in units of reads per
million, required to nucleate a small RNA cluster during de novo cluster
search. Must be an floating point number > 0. Default is 2.}

\item{pad}{integer; initial peaks are merged if they are this distance or
less from each other. Must >= 1, default is 75.}

\item{tidy}{logical; removes unnecessary extra output files when set to TRUE.}
}
\value{
The OS commands generate output into the users desired location, generating
two folders:
\itemize{
\item "1_de_novo_detection" : Stores output from the detection of de novo sRNA-producing loci
\item "2_alignment_results" : Stores output from alignment
}

The first folder stores the novo detection of sRNA-producing loci and
alignment, where the output of each are stored in their respective folders in
the users desired location. The de novo detection of sRNA-producing loci
analyses each sample to identify de novo sRNA-producing loci (ie. sRNA
clusters), and joins these results into a single file called "locifile.txt".
The second folder stores the alignment and clustering results for each sample
to the genome reference along with the file containing the de novo sRNA
clusters. The output of each sample is stored within its own sample folder.

The alignment results generate one folder per sample, where the results
are stored. As default this includes the alignment files (BAM) and
sRNA results (.txt). The sRNA results (.txt) are imported into R for
downstream analysis by utilizing the \code{\link[=RNAimport]{RNAimport()}} function.

The function generates a number of extra files for each sample and are not
required for the downstream analysis. See \code{ShortStack} documentation for more
information (\link{https://github.com/MikeAxtell/ShortStack}). As default
these files are deleted. This is determined by the \code{tidy} argument.
}
\description{
Map sRNAseq FASTQ samples to a given genome assembly using
ShortStack.
}
\details{
This function undertakes our recommended workflow for mapping sRNAseq reads
to a given genome for the analysis of small RNAs in a biological system.
See appendix of vignette for manual pipeline.

The \code{sRNAmapper()} function invokes a number of OS commands, and is dependent
on the installation of \code{ShortStack} (>= 4.0) with Conda.  Please refer to
\link{https://github.com/MikeAxtell/ShortStack} for more information. Please
note that \code{ShortStack} is only compatible with Linux and Mac operating
systems.

The pipeline undertakes de novo detection of sRNA-producing loci and
alignment, where the output of each are stored in their respective folders in
the users desired location. The de novo detection of sRNA-producing loci
analyses each sample to identify de novo sRNA-producing loci (ie. sRNA
clusters), and joins these results into a single file called "locifile.txt".
The alignment step aligns and clusters each sample to the genome reference
along with the file containing the de novo sRNA clusters.
}
\examples{
\dontrun{

samples <- file.path(system.file("extdata",package="mobileRNA"))

output_location <- tempdir()

sRNAmapper(input_files_dir = samples, 
output_dir = output_location, 
genomefile = output_assembly_file,
condaenv = "ShortStack4")
}

}
\references{
ShortStack \link{https://github.com/MikeAxtell/ShortStack}
}
